+++ src/interaction.py
@@ -3070,2175 +3070,6 @@ def processInput(key, charState=None, noAdvanceGame=False, char=None):
         footer.set_text((urwid.AttrSpec("default", "default"), text))
 
 
-
-
-# bad code: there is redundant code from the specific submenus that should be put here
-# bad code: there is spcific code from the selection submenu that should NOT be here
-class SubMenu:
-    """
-    The base class for submenus offering selections
-    """
-
-    def __init__(self, default=None, targetParamName="selection"):
-        """
-        set up basic state
-
-        Parameters:
-            default: the default selection
-            targetParamName: name of the parameter the selection should be stored in
-        """
-
-        self.stealAllKeys = False
-        self.state = None
-        self.options = {}
-        self.selection = None
-        self.selectionIndex = 1
-        self.persistentText = ""
-        self.footerText = "press w / s to move selection up / down, press enter / j / k to select, press esc to exit"
-        self.followUp = None
-        self.done = False
-        self.tag = None
-        self.extraInfo ={}
-
-        self.options = collections.OrderedDict()
-        self.niceOptions = collections.OrderedDict()
-        self.default = default
-        self.targetParamName = targetParamName
-        self.extraDescriptions = {}
-        super().__init__()
-
-        self.escape = False
-
-    def rerender(self):
-        pass
-
-    def callIndirect(self, callback, extraParams=None):
-        """
-        call a callback that is stored in a savable format
-
-        Parameters:
-            callback: the callback to call
-            extraParams: some additional parameters
-        """
-
-        if extraParams is None:
-            extraParams = {}
-        if not isinstance(callback, dict):
-            # bad code: direct function calls are deprecated, but not completely removed
-            callback()
-        else:
-            if "container" not in callback:
-                return
-            container = callback["container"]
-            function = getattr(container, callback["method"])
-
-            if "params" in callback:
-                callback["params"].update(extraParams)
-                function(callback["params"])
-            else:
-                function()
-
-    def getQuestMarkersTile(self, character):
-        return []
-
-    def setOptions(self, query, options):
-        """
-        set the options to select from
-
-        Parameters:
-            query: the text shown for the selection
-            options: the options to choose from
-        """
-
-        # convert options to ordered dict
-
-        self.options = collections.OrderedDict()
-        self.niceOptions = collections.OrderedDict()
-        counter = 1
-        for option in options:
-            self.options[str(counter)] = option[0]
-            self.niceOptions[str(counter)] = option[1]
-            counter += 1
-
-        # set related state
-        self.query = query
-        self.selectionIndex = 1
-        self.lockOptions = True
-        self.selection = None
-        self.origKey = None
-
-    def getSelection(self):
-        """
-        get the selected item
-        """
-
-        return self.selection
-
-    def handleKey(self, key, noRender=False, character=None):
-        """
-        show the options and allow the user to select one
-
-        Parameters:
-            key: the key pressed
-            noRender: flag for skipping the rendering
-        Returns:
-            returns True when done
-        """
-
-        origKey = key
-
-        if not self.options:
-            self.done = True
-            return True
-        if self.done:
-            return True
-
-        # exit submenu
-        if key == "esc":
-            return True
-
-        if key == "a":
-            # convert options to ordered dict
-            oldOptions = self.options
-            oldNiceOptions = self.niceOptions
-
-            self.options = collections.OrderedDict()
-            self.niceOptions = collections.OrderedDict()
-            counter = 1
-            while counter < len(oldOptions):
-                self.options[str(counter)] = oldOptions[str(counter + 1)]
-                self.niceOptions[str(counter)] = oldNiceOptions[str(counter + 1)]
-                counter += 1
-            self.options[str(counter)] = oldOptions[str(1)]
-            self.niceOptions[str(counter)] = oldNiceOptions[str(1)]
-
-        if key == "d":
-            # convert options to ordered dict
-            oldOptions = self.options
-            oldNiceOptions = self.niceOptions
-
-            self.options = collections.OrderedDict()
-            self.niceOptions = collections.OrderedDict()
-            counter = 1
-            self.options[str(counter)] = oldOptions[str(len(oldOptions))]
-            self.niceOptions[str(counter)] = oldNiceOptions[str(len(oldOptions))]
-            counter = 2
-            while counter < len(oldOptions) + 1:
-                self.options[str(counter)] = oldOptions[str(counter - 1)]
-                self.niceOptions[str(counter)] = oldNiceOptions[str(counter - 1)]
-                counter += 1
-
-        # show question
-        out = ["\n"]
-        out.extend([self.query, "\n"])
-
-        # handle the selection of options
-        if not self.lockOptions:
-            # change the marked option
-            if key in (
-                "w",
-                "up",
-            ):
-                self.selectionIndex -= 1
-                if self.selectionIndex == 0:
-                    self.selectionIndex = len(self.options)
-            if key in (
-                "s",
-                "down",
-            ):
-                self.selectionIndex += 1
-                if self.selectionIndex > len(self.options):
-                    self.selectionIndex = 1
-            # select the marked option
-            if key in ["enter", "j", "k", "right"]:
-                # bad code: transforming the key to the shortcut is needlessly complicated
-                if len(self.options.items()):
-                    key = list(self.options.items())[self.selectionIndex - 1][0]
-                else:
-                    self.selection = None
-                    self.done = True
-                    return True
-
-            # select option by shortcut
-            if key in self.options:
-                self.selection = self.options[key]
-                self.options = None
-                if self.followUp:
-                    self.callIndirect(self.followUp,extraParams={self.targetParamName:self.selection,"key":origKey})
-                self.origKey = origKey
-                return True
-        else:
-            self.lockOptions = False
-
-        extraDescription = None
-        if not noRender:
-            # render the options
-            counter = 0
-            for k, v in self.niceOptions.items():
-                counter += 1
-                if counter == self.selectionIndex:
-                    out.extend([" -> ", v, "\n"])
-                    if self.extraDescriptions and self.options[k] in self.extraDescriptions:
-                        extraDescription = self.extraDescriptions[self.options[k]]+"\n\n"
-                else:
-                    out.extend(["    ", v, "\n"])
-
-            if extraDescription:
-                out += extraDescription
-
-            # show the rendered options
-            # bad code: urwid specific code
-            if main:
-                main.set_text(
-                    (
-                        urwid.AttrSpec("default", "default"),
-                        [self.persistentText , "\n\n" , out],
-                    )
-                )
-
-        return False
-
-    # bad code: should either be used everywhere or be removed
-    # bad code: urwid specific code
-    def set_text(self, text):
-        """
-        set text in urwid
-
-        Parameters:
-            text: the text to set
-        """
-
-        main.set_text((urwid.AttrSpec("default", "default"), text))
-
-# bad code: this does nothing the Submenu doesn't do
-class ListActionMenu(SubMenu):
-    """
-    does a simple selection and terminates
-    """
-
-    def __init__(self, options, actions, text="", default=None, targetParamName="selection"):
-        """
-        set up the selection
-
-        Parameters:
-            text: the text to show next to the selection
-            options: the options to select from
-            default: the default value
-            targetParamName: name of the parameter the selection should be stored in
-        """
-
-        self.type = "ListActionMenu"
-        super().__init__(default=default,targetParamName=targetParamName)
-        self.setOptions(text, options)
-        self.actions = actions
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        handles a keypress
-
-        Parameters:
-            key: the key pressed
-            noRender: flag for skipping rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit submenu
-        if key == "esc":
-            self.selection = None
-            if self.followUp:
-                self.callIndirect(self.followUp)
-            return True
-        if not noRender:
-            header.set_text("")
-
-        if key == "a":
-            # convert options to ordered dict
-            oldOptions = self.options
-            oldNiceOptions = self.niceOptions
-
-            self.options = collections.OrderedDict()
-            self.niceOptions = collections.OrderedDict()
-            counter = 1
-            while counter < len(oldOptions):
-                self.options[str(counter)] = oldOptions[str(counter + 1)]
-                self.niceOptions[str(counter)] = oldNiceOptions[str(counter + 1)]
-                counter += 1
-            self.options[str(counter)] = oldOptions[str(1)]
-            self.niceOptions[str(counter)] = oldNiceOptions[str(1)]
-
-        if key == "d":
-            # convert options to ordered dict
-            oldOptions = self.options
-            oldNiceOptions = self.niceOptions
-
-            self.options = collections.OrderedDict()
-            self.niceOptions = collections.OrderedDict()
-            counter = 1
-            self.options[str(counter)] = oldOptions[str(len(oldOptions))]
-            self.niceOptions[str(counter)] = oldNiceOptions[str(len(oldOptions))]
-            counter = 2
-            while counter < len(oldOptions) + 1:
-                self.options[str(counter)] = oldOptions[str(counter - 1)]
-                self.niceOptions[str(counter)] = oldNiceOptions[str(counter - 1)]
-                counter += 1
-
-        if key in (commandChars.autoAdvance, commandChars.advance):
-            if self.default is not None:
-                self.selection = self.default
-            else:
-                self.selection = list(self.options.values())[0]
-            self.options = None
-            if self.followUp:
-                self.followUp()
-            return True
-
-        # show question
-        out = "\n"
-        out += self.query + "\n"
-
-        # change the marked option
-        if key in (
-            "w",
-            "up",
-        ):
-            self.selectionIndex -= 1
-            if self.selectionIndex == 0:
-                self.selectionIndex = len(self.options)
-        if key in (
-            "s",
-            "down",
-        ):
-            self.selectionIndex += 1
-            if self.selectionIndex > len(self.options):
-                self.selectionIndex = 1
-        # select the marked option
-        if key in ["enter", "j", "k", "right"]:
-            key = "default"
-
-        if key in self.actions:
-            option = list(self.options.items())[self.selectionIndex - 1][1]
-            callback = self.actions[key]["callback"]
-            if not callback.get("params"):
-                callback["params"] = {}
-            self.callIndirect(callback,extraParams={self.targetParamName:option})
-            return True
-
-        if not noRender:
-            # render the options
-            counter = 0
-            for k, v in self.niceOptions.items():
-                counter += 1
-                if counter == self.selectionIndex:
-                    out += str(k) + " ->" + str(v) + "\n"
-                else:
-                    out += str(k) + " - " + str(v) + "\n"
-
-            # show the rendered options
-            # bad code: urwid specific code
-            main.set_text(
-                (
-                    urwid.AttrSpec("default", "default"),
-                    self.persistentText + "\n\n" + out,
-                )
-            )
-
-        return False
-
-# bad code: this does nothing the Submenu doesn't do
-class SelectionMenu(SubMenu):
-    """
-    does a simple selection and terminates
-    """
-
-    def __init__(self, text="", options=None, default=None, targetParamName="selection",extraDescriptions=None, selected=None):
-        """
-        set up the selection
-
-        Parameters:
-            text: the text to show next to the selection
-            options: the options to select from
-            default: the default value
-            targetParamName: name of the parameter the selection should be stored in
-        """
-
-        if not options:
-            options = []
-
-        self.type = "SelectionMenu"
-        super().__init__(default=default,targetParamName=targetParamName)
-        self.setOptions(text, options)
-
-        if selected:
-            counter = 0
-            for option in options:
-                counter += 1
-                if option[0] == selected:
-                    self.selectionIndex = counter
-
-        self.extraDescriptions = extraDescriptions
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        handles a keypress
-
-        Parameters:
-            key: the key pressed
-            noRender: flag for skipping rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit submenu
-        if key == "esc":
-            self.selection = None
-            if self.followUp:
-                self.callIndirect(self.followUp,extraParams={self.targetParamName:None})
-            return True
-        if not noRender and header:
-            header.set_text("")
-
-        # let superclass handle the actual selection
-        if not self.getSelection():
-            super().handleKey(key, noRender=noRender, character=character)
-
-        # stop when done
-        return bool(self.getSelection())
-
-class setNPCDutiesMenu(SubMenu):
-    def __init__(self,npc=None):
-        self.npc = npc
-        self.type = "setNPCDutiesMenu"
-
-    def handleKey(self, key, noRender=False, character = None):
-        if self.subMenu:
-            subMenuDone = self.subMenu.handleKey(key, noRender=noRender, character=character)
-            if not subMenuDone:
-                return False
-            key = "~"
-
-        # exit the submenu
-        if key == "esc":
-            return True
-        return None
-
-        # set primary duty
-        # set secondary duty
-        # set tertiary duty
-
-class IdleChatNPCMenu(SubMenu):
-    def __init__(self,npc=None):
-        self.npc = npc
-        self.type = "IdleChatNPCMenu"
-        self.subMenu = None
-        self.infoType = None
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        if self.subMenu:
-            subMenuDone = self.subMenu.handleKey(key, noRender=noRender, character=character)
-            if not subMenuDone:
-                return False
-            key = "~"
-
-        # exit the submenu
-        if key == "esc":
-            return True
-
-        if not self.infoType:
-            if not self.subMenu:
-                options = []
-                options.append(("charInfo","Tell me about yourself."))
-                options.append(("showQuests","What are you doing?"))
-                options.append(("showInventory","What is in your inventory?"))
-                options.append(("showFeelings","How are you feeling?"))
-                options.append(("reset","You are behaving eratically. Get yourself together!"))
-                self.subMenu = SelectionMenu("", options)
-                self.handleKey("~", noRender=noRender, character=character)
-                return False
-            self.instructionType = self.subMenu.selection
-            self.subMenu = None
-
-        if self.instructionType == "charInfo":
-            submenue = src.interaction.CharacterInfoMenu(char=self.npc)
-            character.macroState["submenue"] = submenue
-            submenue.handleKey("~", noRender=noRender,character=character)
-            self.subMenu = None
-            return True
-        if self.instructionType == "showQuests":
-            submenue = src.interaction.QuestMenu(char=self.npc)
-            character.macroState["submenue"] = submenue
-            submenue.handleKey("~", noRender=noRender,character=character)
-            self.subMenu = None
-            return True
-        if self.instructionType == "showInventory":
-            submenue = src.interaction.InventoryMenu(char=self.npc)
-            character.macroState["submenue"] = submenue
-            submenue.handleKey("~", noRender=noRender,character=character)
-            self.subMenu = None
-            return True
-        if self.instructionType == "showFeelings":
-            submenue = src.interaction.InventoryMenu(char=self.npc)
-            character.macroState["submenue"] = submenue
-            submenue.handleKey("~", noRender=noRender,character=character)
-            self.subMenu = None
-            return True
-        if self.instructionType == "reset":
-            for quest in self.npc.quests:
-                quest.fail()
-
-            containerQuest = src.quests.questMap["BeUsefull"]()
-            self.npc.quests.append(containerQuest)
-            containerQuest.assignToCharacter(self.npc)
-            containerQuest.activate()
-            containerQuest.autoSolve = True
-
-            self.npc.timeTaken = 0
-        return True
-
-class InstructSubordinatesMenu(SubMenu):
-    def __init__(self,npcs=None):
-        self.type = "InstructSubordinatesMenu"
-        self.subMenu = None
-        self.instructionType = None
-        self.dutyType = None
-        self.commandType = None
-        self.addFront = True
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        if self.subMenu:
-            subMenuDone = self.subMenu.handleKey(key, noRender=noRender, character=character)
-            if not subMenuDone:
-                return False
-            key = "~"
-
-        # exit the submenu
-        if key == "esc":
-            return True
-
-        if not self.instructionType:
-            if not self.subMenu:
-                options = []
-                options.append(("command selection","select from a list of commands"))
-                options.append(("createQuest","create and issue quest"))
-                self.subMenu = SelectionMenu("how do you want to give the instruction?", options)
-                self.handleKey("~", noRender=noRender, character=character)
-                return False
-            self.instructionType = self.subMenu.selection
-            self.subMenu = None
-
-        if self.instructionType == "createQuest":
-            submenue = src.interaction.AdvancedQuestMenu()
-            submenue.activeChar = character
-            submenue.character = self.npc
-            submenue.state = "questSelection"
-            character.macroState["submenue"] = submenue
-            submenue.handleKey("~", noRender=noRender)
-        if self.instructionType == "command selection":
-            if not self.commandType:
-                if not self.subMenu:
-                    options = []
-                    options.append(("defendTile","defend tile"))
-                    options.append(("attackNorth","attack north"))
-                    options.append(("attackWest","attack west"))
-                    options.append(("attackEast","attack east"))
-                    options.append(("attackSouth","attack south"))
-                    options.append(("scavengeTile","scavenge tile"))
-                    options.append(("clearInventory","clear inventory"))
-                    options.append(("equip","equip"))
-                    options.append(("abortLast","abort last command"))
-                    options.append(("stop","stop what you are doing"))
-                    options.append(("continue","continue working"))
-                    options.append(("wait","wait until further command"))
-                    options.append(("dropAll","drop all your items"))
-                    options.append(("goToMyPosition","go to my position"))
-                    options.append(("beUseful","be useful"))
-                    options.append(("beUsefulHere","be useful here"))
-                    options.append(("doDuty","do duty"))
-                    options.append(("doDutyHere","do duty here"))
-                    self.subMenu = SelectionMenu("what command do you want to give?", options)
-                    self.handleKey("~", noRender=noRender, character=character)
-                    return False
-
-                if self.subMenu.origKey == "k":
-                    self.addFront = False
-
-                self.commandType = self.subMenu.selection
-                self.subMenu = None
-
-            if self.commandType== "stop":
-                self.npc.runCommandString("",clear=True)
-                self.npc.macroState["loop"] = []
-                self.npc.macroState["replay"].clear()
-                if "ifCondition" in self.npc.interactionState:
-                    self.npc.interactionState["ifCondition"].clear()
-                    self.npc.interactionState["ifParam1"].clear()
-                    self.npc.interactionState["ifParam2"].clear()
-                for quest in self.npc.quests[:]:
-                    quest.fail()
-                return True
-            if self.commandType== "continue":
-                self.npc.runCommandString("*",clear=True)
-                return True
-            if self.commandType == "beUseful":
-                for npc in character.subordinates:
-                    quest = src.quests.questMap["BeUsefull"]()
-                    quest.autoSolve = True
-                    npc.assignQuest(quest,active=self.addFront)
-                self.subMenu = None
-                return True
-            if self.commandType == "dropAll":
-                self.npc.runCommandString("10l")
-                return True
-            if self.commandType == "beUsefulHere":
-                quest = src.quests.questMap["BeUsefull"](targetPosition=character.getBigPosition())
-                quest.autoSolve = True
-                self.subMenu = None
-                self.npc.assignQuest(quest,active=True)
-                return True
-            if self.commandType == "goToMyPosition":
-                quest = src.quests.questMap["GoToPosition"](targetPosition=character.getSpacePosition())
-                quest.autoSolve = True
-                self.subMenu = None
-                self.npc.assignQuest(quest,active=True)
-
-                if character.container != self.npc.container:
-                    quest = src.quests.questMap["GoToTile"](targetPosition=character.getBigPosition())
-                    quest.autoSolve = True
-                    self.subMenu = None
-                    self.npc.assignQuest(quest,active=True)
-                return True
-            if self.commandType in ("defendTile",):
-                pos = character.getBigPosition()
-                for npc in character.subordinates:
-                    quest = src.quests.questMap["SecureTile"](toSecure=pos,endWhenCleared=False)
-                    quest.autoSolve = True
-                    npc.assignQuest(quest,active=self.addFront)
-                self.subMenu = None
-
-            if self.commandType in ("abortLast",):
-                for npc in character.subordinates:
-                    if npc.quests:
-                        npc.quests[0].fail()
-                self.subMenu = None
-
-            if self.commandType in ("attackWest","attackEast","attackNorth","attackSouth"):
-                pos = character.getBigPosition()
-                if self.commandType == "attackWest":
-                    pos = (pos[0]-1,pos[1],0)
-                if self.commandType == "attackEast":
-                    pos = (pos[0]+1,pos[1],0)
-                if self.commandType == "attackNorth":
-                    pos = (pos[0],pos[1]-1,0)
-                if self.commandType == "attackSouth":
-                    pos = (pos[0],pos[1]+1,0)
-
-                for npc in character.subordinates:
-                    quest = src.quests.questMap["SecureTile"](toSecure=pos,endWhenCleared=True)
-                    quest.autoSolve = True
-                    npc.assignQuest(quest,active=self.addFront)
-
-                self.subMenu = None
-            if self.commandType in ("scavengeTile",):
-                pos = character.getBigPosition()
-
-                for npc in character.subordinates:
-                    quest = src.quests.questMap["ScavengeTile"](targetPosition=pos)
-                    quest.autoSolve = True
-                    npc.assignQuest(quest,active=self.addFront)
-
-                self.subMenu = None
-            if self.commandType in ("equip",):
-                pos = character.getBigPosition()
-
-                for npc in character.subordinates:
-                    quest = src.quests.questMap["Equip"]()
-                    quest.autoSolve = True
-                    npc.assignQuest(quest,active=True)
-
-                self.subMenu = None
-            if self.commandType in ("clearInventory",):
-                pos = character.getBigPosition()
-
-                for npc in character.subordinates:
-                    quest = src.quests.questMap["ClearInventory"]()
-                    quest.autoSolve = True
-                    npc.assignQuest(quest,active=True)
-
-                self.subMenu = None
-            if self.commandType in ("doDutyHere","doDuty"):
-                if not self.dutyType:
-                    if not self.subMenu:
-                        options = []
-                        options.append(("resource gathering","resource gathering"))
-                        options.append(("machine operation","machine operation"))
-                        options.append(("manufacturing","manufacturing"))
-                        options.append(("trap setting","trap setting"))
-                        options.append(("hauling","hauling"))
-                        options.append(("resource fetching","resource fetching"))
-                        options.append(("cleaning","cleaning"))
-                        options.append(("machine placing","machine placing"))
-                        self.subMenu = SelectionMenu("What duty should be done?", options)
-                        self.handleKey("~", noRender=noRender, character=character)
-                        return False
-                    self.dutyType = self.subMenu.selection
-                    self.subMenu = None
-
-                if self.dutyType:
-                    self.npc.duties = [self.dutyType]
-                    pos = None
-                    if self.commandType == "doDutyHere":
-                        pos = character.getBigPosition()
-                    quest = src.quests.questMap["BeUsefull"](targetPosition=pos,strict=True)
-                    quest.autoSolve = True
-                    self.subMenu = None
-                    self.npc.assignQuest(quest,active=True)
-                    return True
-        return True
-
-class InstructNPCMenu(SubMenu):
-    def __init__(self,npc=None):
-        self.npc = npc
-        self.type = "InstructNPCMenu"
-        self.subMenu = None
-        self.instructionType = None
-        self.dutyType = None
-        self.commandType = None
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        if self.subMenu:
-            subMenuDone = self.subMenu.handleKey(key, noRender=noRender, character=character)
-            if not subMenuDone:
-                return False
-            key = "~"
-
-        # exit the submenu
-        if key == "esc":
-            return True
-
-        if not self.instructionType:
-            if not self.subMenu:
-                options = []
-                options.append(("command selection","select from a list of commands"))
-                options.append(("createQuest","create and issue quest"))
-                self.subMenu = SelectionMenu("how do you want to give the instruction?", options)
-                self.handleKey("~", noRender=noRender, character=character)
-                return False
-            self.instructionType = self.subMenu.selection
-            self.subMenu = None
-
-        if self.instructionType == "createQuest":
-            submenue = src.interaction.AdvancedQuestMenu()
-            submenue.activeChar = character
-            submenue.character = self.npc
-            submenue.state = "questSelection"
-            character.macroState["submenue"] = submenue
-            submenue.handleKey("~", noRender=noRender)
-        if self.instructionType == "command selection":
-            if not self.commandType:
-                if not self.subMenu:
-                    options = []
-                    options.append(("attackNorth","attack north"))
-                    options.append(("attackWest","attack west"))
-                    options.append(("attackEast","attack east"))
-                    options.append(("attackSouth","attack south"))
-                    options.append(("stop","stop what you are doing"))
-                    options.append(("continue","continue working"))
-                    options.append(("wait","wait until further command"))
-                    options.append(("dropAll","drop all your items"))
-                    options.append(("goToMyPosition","go to my position"))
-                    options.append(("beUseful","be useful"))
-                    options.append(("beUsefulHere","be useful here"))
-                    options.append(("doDuty","do duty"))
-                    options.append(("doDutyHere","do duty here"))
-                    self.subMenu = SelectionMenu("what command do you want to give?", options)
-                    self.handleKey("~", noRender=noRender, character=character)
-                    return False
-                self.commandType = self.subMenu.selection
-                self.subMenu = None
-
-            if self.commandType== "stop":
-                self.npc.runCommandString("",clear=True)
-                self.npc.macroState["loop"] = []
-                self.npc.macroState["replay"].clear()
-                if "ifCondition" in self.npc.interactionState:
-                    self.npc.interactionState["ifCondition"].clear()
-                    self.npc.interactionState["ifParam1"].clear()
-                    self.npc.interactionState["ifParam2"].clear()
-                for quest in self.npc.quests[:]:
-                    quest.fail()
-                return True
-            if self.commandType== "continue":
-                self.npc.runCommandString("*",clear=True)
-                return True
-            if self.commandType == "beUseful":
-                quest = src.quests.questMap["BeUsefull"]()
-                quest.autoSolve = True
-                self.subMenu = None
-                self.npc.assignQuest(quest,active=True)
-                return True
-            if self.commandType == "dropAll":
-                self.npc.runCommandString("10l")
-                return True
-            if self.commandType == "beUsefulHere":
-                quest = src.quests.questMap["BeUsefull"](targetPosition=character.getBigPosition())
-                quest.autoSolve = True
-                self.subMenu = None
-                self.npc.assignQuest(quest,active=True)
-                return True
-            if self.commandType == "goToMyPosition":
-                quest = src.quests.questMap["GoToPosition"](targetPosition=character.getSpacePosition())
-                quest.autoSolve = True
-                self.subMenu = None
-                self.npc.assignQuest(quest,active=True)
-
-                if character.container != self.npc.container:
-                    quest = src.quests.questMap["GoToTile"](targetPosition=character.getBigPosition())
-                    quest.autoSolve = True
-                    self.subMenu = None
-                    self.npc.assignQuest(quest,active=True)
-                return True
-            if self.commandType == "attackWest":
-                pos = character.getBigPosition()
-                pos = (pos[0]-1,pos[1],0)
-                quest = src.quests.questMap["SecureTile"](toSecure=pos,endWhenCleared=True)
-                quest.autoSolve = True
-                self.subMenu = None
-                self.npc.assignQuest(quest,active=True)
-            if self.commandType in ("doDutyHere","doDuty"):
-                if not self.dutyType:
-                    if not self.subMenu:
-                        options = []
-                        options.append(("resource gathering","resource gathering"))
-                        options.append(("machine operation","machine operation"))
-                        options.append(("manufacturing","manufacturing"))
-                        options.append(("trap setting","trap setting"))
-                        options.append(("hauling","hauling"))
-                        options.append(("resource fetching","resource fetching"))
-                        options.append(("cleaning","cleaning"))
-                        options.append(("machine placing","machine placing"))
-                        self.subMenu = SelectionMenu("What duty should be done?", options)
-                        self.handleKey("~", noRender=noRender, character=character)
-                        return False
-                    self.dutyType = self.subMenu.selection
-                    self.subMenu = None
-
-                if self.dutyType:
-                    self.npc.duties = [self.dutyType]
-                    pos = None
-                    if self.commandType == "doDutyHere":
-                        pos = character.getBigPosition()
-                    quest = src.quests.questMap["BeUsefull"](targetPosition=pos,strict=True)
-                    quest.autoSolve = True
-                    self.subMenu = None
-                    self.npc.assignQuest(quest,active=True)
-                    return True
-        return True
-
-class NameGhoulMenu(SubMenu):
-    def __init__(self,npc=None):
-        self.npc = npc
-        self.type = "NameGhoulMenu"
-        self.subMenu = None
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        if self.subMenu:
-            subMenuDone = self.subMenu.handleKey(key, noRender=noRender, character=character)
-            if not subMenuDone:
-                return False
-            key = "~"
-
-        # exit the submenu
-        if key == "esc":
-            return True
-
-        if not self.subMenu:
-            self.subMenu = InputMenu("enter the new name for this Ghoul")
-            self.handleKey("~", noRender=noRender, character=character)
-            return False
-        self.npc.name = self.subMenu.text
-        self.subMenu = None
-
-        return True
-
-# bad code: since there is no need to wait for some return this submenue should not wrap around the Chat menu
-# bad code: sub menues should be implemented in the base class
-class ChatPartnerselection(SubMenu):
-    """
-    Spawns a Chat submenu with a player selected character
-    """
-
-    def __init__(self):
-        """
-        initialise internal state
-        """
-
-        self.type = "ChatPartnerselection"
-        self.subMenu = None
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        set up the selection and spawn the chat
-        keystrokes after the setup will be delegated
-
-        Parameters:
-            key: the key pressed
-            noRender: a flag toskip rendering
-        Returns:
-            returns True when done
-        """
-
-        # wrap around the chat menu
-        if self.subMenu:
-            return self.subMenu.handleKey(key, noRender=noRender, character=character)
-
-        # exit the submenu
-        if key == "esc":
-            return True
-
-        # set title
-        header.set_text((urwid.AttrSpec("default", "default"), "\nConversation menu\n"))
-        out = "\n"
-
-        # offer the player the option to select from characters to talk to
-        # bad code: should be done in __init__
-        if not self.options and not self.getSelection():
-            options = []
-            # get characters in room
-            if src.gamestate.gamestate.mainChar.room:
-                for char in src.gamestate.gamestate.mainChar.room.characters:
-                    if char == src.gamestate.gamestate.mainChar:
-                        continue
-                    if char in src.gamestate.gamestate.mainChar.subordinates:
-                        continue
-                    if char.faction != src.gamestate.gamestate.mainChar.faction:
-                        continue
-                    if not (char.xPosition//15 == src.gamestate.gamestate.mainChar.xPosition//15 and char.yPosition//15 == src.gamestate.gamestate.mainChar.yPosition//15):
-                        continue
-                    options.append((char, char.name))
-            # get character on terrain
-            else:
-                for char in src.gamestate.gamestate.mainChar.terrain.characters:
-                    # bad pattern: should only list nearby characters
-                    if char == src.gamestate.gamestate.mainChar:
-                        continue
-                    if char in src.gamestate.gamestate.mainChar.subordinates:
-                        continue
-                    if char.faction != src.gamestate.gamestate.mainChar.faction:
-                        continue
-                    if not (char.xPosition//15 == src.gamestate.gamestate.mainChar.xPosition//15 and char.yPosition//15 == src.gamestate.gamestate.mainChar.yPosition//15):
-                        continue
-
-                    if char.rank and src.gamestate.gamestate.mainChar.rank:
-                        if char.rank < src.gamestate.gamestate.mainChar.rank:
-                            options.append((char, char.name + " (outranks you)"))
-                        elif char.rank == src.gamestate.gamestate.mainChar.rank:
-                            options.append((char, char.name + " (same rank)"))
-                    else:
-                        options.append((char, char.name))
-
-            for char in src.gamestate.gamestate.mainChar.subordinates:
-                options.insert(0,(char, char.name))
-
-            self.setOptions("talk with whom?", options)
-
-        # delegate the actual selection to the super class
-        if not self.getSelection():
-            super().handleKey(key, noRender=noRender, character=character)
-
-        # spawn the chat submenu
-        if self.getSelection():
-            self.subMenu = src.chats.ChatMenu(self.selection)
-            self.subMenu.handleKey(key, noRender=noRender, character=character)
-            return None
-
-        # wait for input
-        else:
-            return False
-
-class DebugMenu(SubMenu):
-    """
-    menu offering minimal debug ability
-    (actually does nothing)
-    """
-
-    def __init__(self):
-        """
-        initialise internal state
-        """
-
-        self.type = "DebugMenu"
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show some debug output
-        (actually does nothing)
-
-        Parameters:
-            key: the key pressed
-            noRender: a flag toskip rendering
-        Returns:
-            returns True when done
-        """
-        self.persistentText = ["debug"]
-
-        # exit submenu
-        if key == "esc":
-            return True
-        return None
-
-class QuestMenu(SubMenu):
-    """
-    show the quests for a character and allow player interaction
-    """
-
-    def __init__(self, char=None):
-        """
-        initialise internal state
-
-        Parameters:
-            char: the character to show the quests for
-        """
-
-        self.type = "QuestMenu"
-        self.lockOptions = True
-        if not char:
-            char = src.gamestate.gamestate.mainChar
-        self.char = char
-        self.offsetX = 0
-        self.questCursor = [0]
-        self.sidebared = False
-        super().__init__()
-
-    def render(self, char):
-        return renderQuests(char=self.char, asList=True, questCursor=self.questCursor,sidebared=self.sidebared)
-
-    # overrides the superclasses method completely
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show a questlist and handle interactions
-
-        Parameters:
-            key: the key pressed
-            noRender: a flag toskip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit submenu
-        if key == "esc":
-            return True
-        if key in ("ESC","lESC",):
-            self.char.rememberedMenu.append(self)
-            self.sidebared = True
-            return True
-        if key in ("rESC",):
-            self.char.rememberedMenu2.append(self)
-            self.sidebared = True
-            return True
-
-        # move the marker that marks the selected quest
-        if key == "w" and self.questCursor[0] > 0:
-            self.questCursor[0] -= 1
-        if key == "s" and self.questCursor[0] < len(character.quests)-1:
-            self.questCursor[0] += 1
-        if key == "d":
-            baseList = self.char.quests
-            failed = False
-            for index in self.questCursor:
-                quest = baseList[index]
-                try:
-                    baseList = quest.subQuests
-                    if len(baseList) < 1:
-                        failed = True
-                except:
-                    baseList = None
-                    failed = True
-            if not failed:
-                self.questCursor.append(0)
-        if key == "a" and len(self.questCursor) > 1:
-            self.questCursor.pop()
-
-        # make the selected quest active
-        if key == "j" and self.questCursor[0]:
-            quest = self.char.quests[self.questCursor[0]]
-            self.char.quests.remove(quest)
-            self.char.quests.insert(0, quest)
-            self.char.setPathToQuest(quest)
-            self.questCursor[0] = 0
-            self.char.runCommandString(["esc"])
-        if key == "K":
-            quest = None
-            baseList = self.char.quests
-            for index in self.questCursor:
-                quest = baseList[index]
-                baseList = quest.subQuests
-            quest.autoSolve = True
-            self.char.runCommandString(["esc"])
-        if key == "r":
-            quest = None
-            baseList = self.char.quests
-            for index in self.questCursor:
-                quest = baseList[index]
-                baseList = quest.subQuests
-            quest.generateSubquests(self.char)
-        if key == "R":
-            baseList = self.char.quests
-            for index in self.questCursor:
-                quest = baseList[index]
-                baseList = quest.subQuests
-            quest.clearSubQuests()
-            quest.generateSubquests(self.char)
-        if key == "x":
-            baseList = self.char.quests
-            for index in self.questCursor:
-                quest = baseList[index]
-                try:
-                    baseList = quest.subQuests
-                except:
-                    baseList = None
-            quest.fail()
-        if key == "X":
-            baseList = self.char.quests
-            for index in self.questCursor:
-                quest = baseList[index]
-                try:
-                    baseList = quest.subQuests
-                except:
-                    baseList = None
-            quest.clearSubQuests()
-
-        # render the quests
-        addition = ""
-        if self.char == src.gamestate.gamestate.mainChar:
-            addition = " (you)"
-        header.set_text(
-            (
-                urwid.AttrSpec("default", "default"),
-                "\nquest overview for "
-                + self.char.name
-                + ""
-                + addition
-                + "\n\n",
-            )
-        )
-        self.persistentText = []
-        self.persistentText.append(
-            self.render(self.char)
-        )
-
-        self.lockOptions = False
-
-        # add interaction instructions
-        self.persistentText.extend(
-            [
-                "\n",
-                "* press esc to close this menu\n",
-                "* press wasd to select quest\n",
-                "* press j to make selected quest the active quest\n",
-                "* press x to delete selected quest\n",
-                "* press X to delete sub quests\n",
-                "* press r to generate sub quests\n",
-                "* press R to regenerate sub quests\n",
-                "* press k to check if that quest has been completed\n",
-                "* press K to mark the selected quest for auto completion\n",
-                "\n",
-            ]
-        )
-
-        # flatten the mix of strings and urwid format so that it is less recursive to workaround an urwid bug
-        # bad code: should be elsewhere
-        def flatten(pseudotext):
-            newList = []
-            for item in pseudotext:
-                if isinstance(item, list):
-                    for subitem in flatten(item):
-                        newList.append(subitem)
-                elif isinstance(item, tuple):
-                    newList.append((item[0], flatten(item[1])))
-                else:
-                    newList.append(item)
-            return newList
-
-        self.persistentText = flatten(self.persistentText)
-
-        # show rendered quests via urwid
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        return False
-
-# bad code: should be abstracted
-# bad code: uses global functions to render
-class InventoryMenu(SubMenu):
-    """
-    shows and interacts with a characters inventory
-    """
-
-    type = "InventoryMenu"
-
-    def __init__(self, char=None):
-        """
-        initialise the internal state
-
-        Parameters:
-            char: the character that owns the inventory that should be shown
-        """
-
-        self.subMenu = None
-        self.skipKeypress = False
-        self.activate = False
-        self.drop = False
-        self.char = char
-        self.sidebared = False
-        self.cursor = 0
-        super().__init__()
-        self.footerText = "press j to activate, press l to drop, press esc to exit"
-
-    def render(self,char=None):
-        try:
-            self.cursor
-        except:
-            self.cursor = 0
-        return renderInventory(character=char,sidebared=self.sidebared,cursor=self.cursor)
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the inventory
-
-        Parameters:
-            key: the key pressed
-            noRender: a flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        if self.subMenu:
-            self.subMenu.handleKey(key, noRender=noRender, character=character)
-            if self.drop:
-                direction = self.subMenu.keyPressed
-                if direction == "w":
-                    pos = (self.char.xPosition, self.char.yPosition - 1, self.char.zPosition)
-                elif direction == "s":
-                    pos = (self.char.xPosition, self.char.yPosition + 1, self.char.zPosition)
-                elif direction == "d":
-                    pos = (self.char.xPosition + 1, self.char.yPosition + 0, self.char.zPosition)
-                elif direction == "a":
-                    pos = (self.char.xPosition - 1, self.char.yPosition + 0, self.char.zPosition)
-                else:
-                    pos = (self.char.xPosition, self.char.yPosition, self.char.zPosition)
-
-                item = self.char.inventory[self.cursor]
-                self.char.addMessage(f"you drop a {item.type}")
-                self.char.drop(self.char.inventory[self.cursor],pos)
-                self.char.timeTaken += self.char.movementSpeed
-
-                self.drop = False
-                self.subMenu = None
-                return True
-            self.subMenu = None
-            self.skipKeypress = True
-            return False
-
-        if self.skipKeypress:
-            self.skipKeypress = False
-        else:
-            # exit the submenu
-            if key == "esc":
-                return True
-            if key in ("ESC","lESC",):
-                self.char.rememberedMenu.append(self)
-                self.sidebared = True
-                return True
-            if key in ("rESC",):
-                self.char.rememberedMenu2.append(self)
-                self.sidebared = True
-                return True
-
-            # do activation
-            if key == "j":
-                item = self.char.inventory[self.cursor]
-                self.char.addMessage(f"you activate the {item.type}")
-                item.apply(self.char)
-                self.char.timeTaken += self.char.movementSpeed
-
-            # do drop
-            if key == "l":
-                item = self.char.inventory[self.cursor]
-                self.char.addMessage(f"you drop a {item.type}")
-                self.char.drop(self.char.inventory[self.cursor])
-                self.char.timeTaken += self.char.movementSpeed
-
-            # do drop
-            if key == "L":
-                self.subMenu = OneKeystrokeMenu("Drop where?\n\n w - north\n s - south\n a - left\n d - right")
-                self.subMenu.handleKey("~", noRender=noRender,character=character)
-                self.drop = True
-                return False
-
-            # handle cursor movement
-            if key == "w":
-                self.cursor -= 1
-            if key == "s":
-                self.cursor += 1
-
-            # handle out of bounds cursor
-            if self.cursor > len(self.char.inventory)-1:
-                self.cursor = 0
-            if self.cursor < 0:
-                self.cursor = len(self.char.inventory)-1
-
-        if not noRender:
-            header.set_text(
-                (urwid.AttrSpec("default", "default"), "\ninventory overview\n\n")
-            )
-
-            # bad code: uses global function
-            self.persistentText = (
-                urwid.AttrSpec("default", "default"),
-                self.render(self.char),
-            )
-
-            # show the render
-            main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        return False
-
-class InputMenu(SubMenu):
-    """
-    menu to get a string input from the user
-    """
-
-    type = "InputMenu"
-
-    def __init__(self, query="", ignoreFirst=False, targetParamName="text",stealAllKeys=False):
-        """
-        initialise internal state
-
-        Parameters:
-            query: the text to be shown along with the input prompt
-            ignoreFirst: flag to ignore first keypress
-        """
-
-        self.query = query
-        self.text = ""
-        super().__init__()
-        self.footerText = "enter the text press enter to confirm"
-        self.firstHit = True
-        self.ignoreFirst = ignoreFirst
-        self.escape = False
-        self.position = 0
-        self.targetParamName = targetParamName
-        self.stealAllKeys = stealAllKeys
-        self.done = False
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        gather the input keystrokes
-
-        Parameters:
-            key: the key pressed
-            noRender: a flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        if key == "enter" and not self.escape:
-            if self.followUp:
-                self.callIndirect(self.followUp,extraParams={self.targetParamName:self.text})
-            self.done = True
-            return True
-
-        if self.ignoreFirst and self.firstHit:
-            pass
-        elif key == "\\" and not self.escape:
-            self.escape = True
-        elif key == "backspace" and not self.escape:
-            if self.position:
-                self.text = (
-                    self.text[0: self.position - 1] + self.text[self.position:]
-                )
-                self.position -= 1
-        elif key == "delete" and not self.escape:
-            if self.position < len(self.text):
-                self.text = (
-                    self.text[0: self.position] + self.text[self.position + 1:]
-                )
-        elif key == "~":
-            pass
-        elif key == "+":
-            return None
-        elif key == "*":
-            return None
-        elif key == "left":
-            self.position -= 1
-        elif key == "right":
-            self.position += 1
-        else:
-            if key == "enter":
-                key = "\n"
-            if len(self.text):
-                self.text = (
-                    self.text[0: self.position] + key + self.text[self.position:]
-                )
-            else:
-                self.text = key
-            self.position += 1
-            self.escape = False
-
-        if len(self.text):
-            text = self.text[0: self.position] + "█" + self.text[self.position:]
-        else:
-            text = "█"
-
-        if not noRender:
-            header.set_text((urwid.AttrSpec("default", "default"), "\ntext input\n\n"))
-            footer.set_text((urwid.AttrSpec("default", "default"), "\ntext input\n\n"))
-
-            self.persistentText = (
-                urwid.AttrSpec("default", "default"),
-                "\n" + self.query + "\n\n" + text,
-            )
-
-            # show the render
-            main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        if self.firstHit:
-            self.firstHit = False
-
-        return False
-
-class MessagesMenu(SubMenu):
-    def render(self,char):
-        if self.scrollIndex:
-            return "\n".join(reversed(char.messages[-46-self.scrollIndex:-self.scrollIndex]))
-        else:
-            return "\n".join(reversed(char.messages[-46:]))
-
-    type = "MessagesMenu"
-
-    def __init__(self, char=None):
-        self.char = char
-        self.scrollIndex = 0
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the attributes and ignore keystrokes
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key == "a" and self.scrollIndex > 0:
-            self.scrollIndex -= 1
-        if key == "d":
-            self.scrollIndex += 1
-        if key == "esc":
-            character.changed("closedMessages")
-            return True
-        if key in ("ESC","lESC",):
-            self.char.rememberedMenu.append(self)
-            return True
-        if key in ("rESC",):
-            self.char.rememberedMenu2.append(self)
-            return True
-
-        char = self.char
-
-        text = f"press a/d to scroll\noldest message on top - skipping {self.scrollIndex} messages\n\n"+self.render(char)
-
-        # show info
-        header.set_text((urwid.AttrSpec("default", "default"), "messages"))
-        main.set_text((urwid.AttrSpec("default", "default"), [text]))
-        header.set_text((urwid.AttrSpec("default", "default"), ""))
-        return None
-
-# bad code: should be abstracted
-# bad code: uses global function to render
-class CombatInfoMenu(SubMenu):
-    """
-    menu to show the players attributes
-    """
-
-    type = "CombatInfoMenu"
-
-    def __init__(self, char=None):
-        self.char = char
-        super().__init__()
-
-    def render(self,char):
-        if char.dead:
-            return ""
-
-        text = ""
-
-        text += "you: \n\n"
-        text += "name:        %s\n" % char.name
-        text += "health:      %s\n" % char.health
-        text += "exhaustion:  %s\n" % char.exhaustion
-        text += "timeTaken:   %f\n" % char.timeTaken
-
-        text += """
-
-nearby enemies:
-
-"""
-
-        enemies = char.getNearbyEnemies()
-        for enemy in enemies:
-            text += "-------------  \n"
-            text += "name:        %s\n" % enemy.name
-            text += "health:     %s\n" % enemy.health
-            text += "exhaustion:  %s\n" % enemy.exhaustion
-            timeTaken = enemy.timeTaken
-            if timeTaken > 1:
-                timeTaken -= 1
-            text += f"timeTaken:   {timeTaken:f}\n"
-
-        text += """
-
-subordinates:
-
-"""
-        for ally in char.subordinates:
-            text += "-------------  \n"
-            text += "name:        %s\n" % ally.name
-            text += "health:     %s\n" % ally.health
-            text += "exhaustion:  %s\n" % ally.exhaustion
-            timeTaken = ally.timeTaken
-            if timeTaken > 1:
-                timeTaken -= 1
-            text += f"timeTaken:   {timeTaken:f}\n"
-
-        return text
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the attributes and ignore keystrokes
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key == "esc":
-            return True
-        if key in ("ESC","lESC",):
-            self.char.rememberedMenu.append(self)
-            return True
-        if key in ("rESC",):
-            self.char.rememberedMenu2.append(self)
-            return True
-
-        char = self.char
-
-        text = self.render(char)
-
-        # show info
-        header.set_text((urwid.AttrSpec("default", "default"), "\ncharacter overview"))
-        main.set_text((urwid.AttrSpec("default", "default"), [text]))
-        header.set_text((urwid.AttrSpec("default", "default"), ""))
-        return None
-
-
-# bad code: should be abstracted
-# bad code: uses global function to render
-class CharacterInfoMenu(SubMenu):
-    """
-    menu to show the players attributes
-    """
-
-    type = "CharacterInfoMenu"
-
-    def __init__(self, char=None):
-        self.char = char
-        super().__init__()
-
-    def render(self,char):
-        if char.dead:
-            return ""
-
-        text = ""
-
-        armorValue = None
-        if char.armor:
-            armorValue = char.armor.armorValue
-        weaponBaseDamage = None
-        if char.weapon:
-            weaponBaseDamage = char.weapon.baseDamage
-
-        text += "name:       %s\n" % char.name
-        text += "\n"
-        text += "\n"
-        text += "health:     %s" % char.health + "\n"
-        text += "max health: %s" % char.maxHealth + "\n"
-        text += "exhaustion: %s" % char.exhaustion + "\n"
-        text += "\n"
-        text += "baseDamage: %s\n" % char.baseDamage
-        text += "weapon:     %s\n" % weaponBaseDamage
-        text += "armor:      %s\n" % armorValue
-        text += "faction:    %s\n" % char.faction
-        text += "time taken: %s" % char.timeTaken + "\n"
-
-        if hasattr(char,"rank"):
-            text += "rank:       %s\n" % char.rank
-        if hasattr(char,"superior"):
-            text += "superior:   %s\n" % char.superior
-        text += "reputation: %s\n" % char.reputation
-        flaskInfo = "-"
-        if char.flask:
-            flaskInfo = str(char.flask.uses)+" flask charges"
-        text += f"satiation:  {char.satiation} ({flaskInfo})\n"
-
-        text += "\n"
-        text += f"movementSpeed:  {char.movementSpeed}\n"
-        text += f"attackSpeed:    {char.attackSpeed}\n"
-        text += "\n"
-        for jobOrder in char.jobOrders:
-            text += str(jobOrder.taskName)
-            text += ": %s \n" % json.dumps(jobOrder.tasks)#,indent=4)
-        text += "\n"
-        text += "lastJobOrder: %s\n" % char.lastJobOrder
-        text += "skills: %s\n" % char.skills
-        if len(char.duties) > 5:
-            text += "duties: %s\n" % ",\n".join(char.duties)
-        else:
-            text += "duties: %s\n" % char.duties
-        text += "numAttackedWithoutResponse: %s\n" % char.numAttackedWithoutResponse
-        text += f"position: {char.getSpacePosition()}\n"
-        text += f"big position: {char.getBigPosition()}\n"
-        text += f"terrain position: {char.getTerrainPosition()}\n"
-        text += f"grievances: {char.grievances}\n"
-        text += f"registers: {char.registers}\n"
-        text += f"terrainName: %s\n" % char.getTerrain().tag
-        text += f"disableCommandsOnPlus: %s\n" % char.disableCommandsOnPlus
-        text += f"autoExpandQuests: %s\n" % char.autoExpandQuests
-        text += f"autoExpandQuests2: %s\n" % char.autoExpandQuests2
-
-        return text
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the attributes and ignore keystrokes
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key == "esc":
-            return True
-        if key in ("ESC","lESC",):
-            self.char.rememberedMenu.append(self)
-            return True
-        if key in ("rESC",):
-            self.char.rememberedMenu2.append(self)
-            return True
-
-        char = self.char
-
-        text = self.render(char)
-
-        # show info
-        header.set_text((urwid.AttrSpec("default", "default"), "\ncharacter overview"))
-        main.set_text((urwid.AttrSpec("default", "default"), [text]))
-        header.set_text((urwid.AttrSpec("default", "default"), ""))
-        return None
-
-class CreateQuestMenu(SubMenu):
-    type = "CreateQuestMenu"
-
-    def __init__(self, questType=None, assignTo=None, activeChar=None):
-        self.requiredParams = None
-        self.questParams = {}
-        self.questType = questType
-        self.quest = None
-        self.submenu = None
-        self.assignTo = assignTo
-        super().__init__()
-        self.stealAllKeys = False
-        self.parameterName = None
-        self.parameterValue = None
-        self.activeChar = activeChar
-
-    def handleKey(self, key, noRender=False, character = None):
-        # exit submenu
-        if key == "esc":
-            return True
-
-        if not self.quest:
-            self.quest = src.quests.questMap[self.questType]()
-
-        if self.submenu:
-            if not self.submenu.handleKey(key, noRender=noRender, character=character):
-                return False
-            param = self.requiredParams.pop()
-
-            rawParameter = self.submenu.text
-            if param["type"] == "int":
-                self.questParams[param["name"]] = int(rawParameter)
-            elif param["type"] == "string":
-                self.questParams[param["name"]] = rawParameter
-            elif param["type"] == "coordinate":
-                if rawParameter == ".":
-                    self.questParams[param["name"]] = character.getBigPosition()
-                else:
-                    self.questParams[param["name"]] = (int(rawParameter.split(",")[0]),int(rawParameter.split(",")[1]),0)
-            self.submenu = None
-
-        if self.requiredParams is None:
-            self.requiredParams = self.quest.getRequiredParameters()
-
-        if self.requiredParams and not self.submenu:
-            param = self.requiredParams[-1]
-            description = "set param: "
-            if param["type"] == "coordinate":
-                description += str(character.getBigPosition())
-            self.submenu = src.interaction.InputMenu(f"{description}{param}")
-            self.submenu.handleKey("~", noRender=noRender, character=character)
-            self.stealAllKeys = True
-            return False
-
-        if not self.requiredParams and key == " ":
-            for char in self.assignTo:
-                if char is None or char.dead:
-                    continue
-
-                quest = src.quests.questMap[self.questType]()
-                quest.setParameters(self.questParams)
-                foundQuest = None
-                for targetQuest in char.quests:
-                    if targetQuest.type == "Serve":
-                        foundQuest = targetQuest
-                        break
-
-                if not foundQuest:
-                    char.assignQuest(quest,active=True)
-                else:
-                    foundQuest.addQuest(quest)
-                quest.activate()
-                quest.assignToCharacter(char)
-                if char == self.activeChar:
-                    quest.selfAssigned = True
-                char.showGotCommand = True
-            self.activeChar.showGaveCommand = True
-            return True
-
-        self.optionalParams = self.quest.getOptionalParameters()
-
-        if key not in ("enter","~"):
-            self.stealAllKeys = True
-            if self.parameterName is None:
-                self.parameterName = ""
-
-            if self.parameterValue is None:
-                if key == ":":
-                    self.parameterValue = ""
-                elif key == "backspace":
-                    if len(self.parameterName):
-                        self.parameterName = self.parameterName[:-1]
-                else:
-                    self.parameterName += key
-            else:
-                if key == ";":
-                    value = None
-                    if self.parameterValue == "None":
-                        value = None
-                    else:
-                        value = int(self.parameterValue)
-
-                    self.questParams[self.parameterName] = value
-                    self.parameterName = None
-                    self.parameterValue = None
-                elif key == "backspace":
-                    if len(self.parameterValue):
-                        self.parameterValue = self.parameterValue[:-1]
-                else:
-                    self.parameterValue += key
-
-        # start rendering
-        if not noRender:
-            header.set_text((urwid.AttrSpec("default", "default"), "\ncreate Quest\n"))
-            # show rendered text via urwid
-            main.set_text((urwid.AttrSpec("default", "default"), "type: {}\n\nparameters: \n\n{}\n\ncurrent parameter: \n\n{} : {}\n\noptional parameters: \n\n{}\n\npress space to confirm".format(self.questType,self.questParams,self.parameterName,self.parameterValue,self.optionalParams)))
-        return False
-
-class AdvancedQuestMenu(SubMenu):
-    """
-    player interaction for delegating a quest
-    """
-
-    type = "AdvancedQuestMenu"
-
-    def __init__(self,activeChar=None):
-        """
-        set up internal state
-        """
-
-        self.character = None
-        self.quest = None
-        self.questParams = {}
-        self.activeChar = activeChar
-        super().__init__()
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        gather the quests parameters and assign the quest
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit submenu
-        if key == "esc":
-            return True
-
-        # start rendering
-        if not noRender:
-            header.set_text(
-                (urwid.AttrSpec("default", "default"), "\nadvanced Quest management\n")
-            )
-            out = "\n"
-            #if self.character:
-            #    out += "character: " + str(self.character.name) + "\n"
-            if self.quest:
-                out += "quest: " + str(self.quest) + "\n"
-            out += "\n"
-
-        # let the player select the character to assign the quest to
-        if self.state is None:
-            self.state = "participantSelection"
-        if self.state == "participantSelection":
-
-            if key == "S":
-                self.state = "questSelection"
-                self.character = "ALL"
-                self.selection = None
-                self.lockOptions = True
-                self.options = None
-            else:
-                # set up the options
-                if not self.options and not self.getSelection():
-
-                    # add the active player as target
-                    options = [(
-                        self.activeChar,
-                        self.activeChar.name+ " (you)",
-                    )]
-
-                    # add the main players subordinates as target
-                    for char in self.activeChar.subordinates:
-                        if char is None:
-                            continue
-                        options.append((char, char.name))
-                    self.setOptions("whom to give the order to: \n(press S for all subordinates)", options)
-
-                # let the superclass handle the actual selection
-                if not self.getSelection():
-                    super().handleKey(key, noRender=noRender, character=character)
-
-                # store the character to assign the quest to
-                if self.getSelection():
-                    self.state = "questSelection"
-                    self.character = self.selection
-                    self.selection = None
-                    self.lockOptions = True
-                else:
-                    return False
-
-        # let the player select the type of quest to create
-        if self.state == "questSelection":
-
-            if key == "N":
-                self.state = "questByName"
-                self.selection = ""
-                key = "~"
-            else:
-                # add quests to select from
-                if not self.options and not self.getSelection():
-                    options = []
-                    """
-                    for key, value in src.quests.questMap.items():
-
-                        # show only quests the character has done
-                        if key not in self.activeChar.questsDone:
-                            continue
-
-                        # do not show naive quests
-                        if key.startswith("Naive"):
-                            continue
-
-                        options.append((value.type, key))
-                    """
-                    options.append(("GoHome", "GoHome"))
-                    options.append(("SecureTile", "SecureTile"))
-                    options.append(("ClearInventory", "ClearInventory"))
-                    options.append(("BeUsefull", "BeUsefull"))
-                    options.append(("BeUsefullOnTile", "BeUsefullOnTile"))
-                    options.append(("Equip", "Equip"))
-                    options.append(("Eat", "Eat"))
-                    options.append(("FillFlask", "FillFlask"))
-                    self.setOptions("what type of quest: (press N for quest by name)", options)
-
-                # let the superclass handle the actual selection
-                if not self.getSelection():
-                    super().handleKey(key, noRender=noRender, character=character)
-
-                # store the type of quest to create
-                if self.getSelection():
-                    self.state = "parameter selection"
-                    self.quest = self.selection
-                    self.selection = None
-                    self.lockOptions = True
-                    self.questParams = {}
-                else:
-                    return False
-
-        if self.state == "questByName":
-            if key == "enter":
-                self.state = "parameter selection"
-                if self.selection not in src.quests.questMap:
-                    return True
-                self.quest = self.selection
-                self.selection = None
-                self.lockOptions = True
-                self.questParams = {}
-            else:
-                if key == "~":
-                    pass
-                elif key == "backspace":
-                    if len(self.selection) > 0:
-                        self.selection = self.selection[:-1]
-                else:
-                    self.selection += key
-
-                if not noRender:
-                    main.set_text((urwid.AttrSpec("default", "default"), self.selection+"\n\n%s"%(self.activeChar.questsDone)))
-                return False
-
-        # let the player select the parameters for the quest
-        if self.state == "parameter selection":
-            if self.quest == "EnterRoomQuestMeta":
-
-                # set up the options
-                if not self.options and not self.getSelection():
-
-                    # add a list of of rooms
-                    options = []
-                    for room in src.gamestate.gamestate.terrain.rooms:
-                        # do not show unimportant rooms
-                        if isinstance(room, src.rooms.CpuWasterRoom | src.rooms.MechArmor):
-                            continue
-                        options.append((room, room.name))
-                    self.setOptions("select the room:", options)
-
-                # let the superclass handle the actual selection
-                if not self.getSelection():
-                    super().handleKey(key, noRender=noRender, character=character)
-
-                # store the parameter
-                if self.getSelection():
-                    self.questParams["room"] = self.selection
-                    self.state = "confirm"
-                    self.selection = None
-                    self.lockOptions = True
-                else:
-                    return False
-
-            elif self.quest == "StoreCargo":
-
-                # set up the options for selecting the cargo room
-                if "cargoRoom" not in self.questParams:
-                    if not self.options and not self.getSelection():
-                        # add a list of of rooms
-                        options = []
-                        for room in src.gamestate.gamestate.terrain.rooms:
-                            # show only cargo rooms
-                            if not isinstance(room, src.rooms.CargoRoom):
-                                continue
-                            options.append((room, room.name))
-                        self.setOptions("select the room:", options)
-
-                    # let the superclass handle the actual selection
-                    if not self.getSelection():
-                        super().handleKey(key, noRender=noRender, character=character)
-
-                    # store the parameter
-                    if self.getSelection():
-                        self.questParams["cargoRoom"] = self.selection
-                        self.selection = None
-                        self.lockOptions = True
-                    else:
-                        return False
-                else:
-                    # set up the options for selecting the storage room
-                    if not self.options and not self.getSelection():
-                        # add a list of of rooms
-                        options = []
-                        for room in src.gamestate.gamestate.terrain.rooms:
-                            # show only storage rooms
-                            if not isinstance(room, src.rooms.StorageRoom):
-                                continue
-                            options.append((room, room.name))
-                        self.setOptions("select the room:", options)
-
-                    # let the superclass handle the actual selection
-                    if not self.getSelection():
-                        super().handleKey(key, noRender=noRender, character=character)
-
-                    # store the parameter
-                    if self.getSelection():
-                        self.questParams["storageRoom"] = self.selection
-                        self.state = "confirm"
-                        self.selection = None
-                        self.lockOptions = True
-                    else:
-                        return False
-            elif self.quest:
-                if self.character == "ALL":
-                    self.activeChar.macroState["submenue"] = CreateQuestMenu(self.quest, self.activeChar.subordinates, self.activeChar)
-                else:
-                    self.activeChar.macroState["submenue"] = CreateQuestMenu(self.quest, [self.character], self.activeChar)
-                return False
-            else:
-                # skip parameter selection
-                self.state = "confirm"
-
-        # get confirmation and assign quest
-        if self.state == "confirm":
-
-            # set the options for confirming the selection
-            if not self.options and not self.getSelection():
-                options = [("yes", "yes"), ("no", "no")]
-                if self.quest == src.quests.EnterRoomQuestMeta:
-                    self.setOptions(
-                        "you chose the following parameters:\nroom: "
-                        + str(self.questParams)
-                        + "\n\nDo you confirm?",
-                        options,
-                    )
-                else:
-                    self.setOptions("Do you confirm?", options)
-
-            # let the superclass handle the actual selection
-            if not self.getSelection():
-                super().handleKey(key, noRender=noRender, character=character)
-
-            if self.getSelection():
-                # instantiate quest
-                # bad code: repetitive code
-                if self.selection == "yes":
-                    if self.quest == src.quests.MoveQuestMeta:
-                        questInstance = self.quest(
-                            src.gamestate.gamestate.mainChar.room, 2, 2
-                        )
-                    elif self.quest == src.quests.ActivateQuestMeta:
-                        questInstance = self.quest(
-                            src.gamestate.gamestate.terrain.tutorialMachineRoom.furnaces[
-                                0
-                            ]
-                        )
-                    elif self.quest == src.quests.EnterRoomQuestMeta:
-                        questInstance = self.quest(self.questParams["room"])
-                    elif self.quest == src.quests.FireFurnaceMeta:
-                        questInstance = self.quest(
-                            src.gamestate.gamestate.terrain.tutorialMachineRoom.furnaces[
-                                0
-                            ]
-                        )
-                    elif self.quest == src.quests.WaitQuest:
-                        questInstance = self.quest()
-                    elif self.quest == src.quests.LeaveRoomQuest:
-                        questInstance = self.quest(self.character.room)
-                    elif self.quest == src.quests.ClearRubble:
-                        questInstance = self.quest()
-                    elif self.quest == src.quests.RoomDuty:
-                        questInstance = self.quest()
-                    elif self.quest == src.quests.ConstructRoom:
-                        for room in src.gamestate.gamestate.terrain.rooms:
-                            if isinstance(room, src.rooms.ConstructionSite):
-                                constructionSite = room
-                                break
-                        questInstance = self.quest(
-                            constructionSite,
-                            src.gamestate.gamestate.terrain.tutorialStorageRooms,
-                        )
-                    elif self.quest == src.quests.StoreCargo:
-                        for room in src.gamestate.gamestate.terrain.rooms:
-                            if isinstance(room, src.rooms.StorageRoom):
-                                storageRoom = room
-                        questInstance = self.quest(
-                            self.questParams["cargoRoom"],
-                            self.questParams["storageRoom"],
-                        )
-                    elif self.quest == src.quests.MoveToStorage:
-                        questInstance = self.quest(
-                            [
-                                src.gamestate.gamestate.terrain.tutorialLab.itemByCoordinates[
-                                    (1, 9)
-                                ][
-                                    0
-                                ],
-                                src.gamestate.gamestate.terrain.tutorialLab.itemByCoordinates[
-                                    (2, 9)
-                                ][
-                                    0
-                                ],
-                            ],
-                            src.gamestate.gamestate.terrain.tutorialStorageRooms[1],
-                        )
-                    elif self.quest == "special_furnace":
-                        questInstance = src.quests.KeepFurnaceFiredMeta(
-                            self.character.room.furnaces[0]
-                        )
-                    else:
-                        questInstance = self.quest()
-
-                    # assign the quest
-
-                    self.character.assignQuest(questInstance, active=True)
-
-                    self.state = "done"
-
-                # reset progress
-                else:
-                    self.state = "questSelection"
-
-                self.selection = None
-                self.lockOptions = False
-            else:
-                return False
-
-        # close submenu
-        if self.state == "done":
-            if self.lockOptions:
-                self.lockOptions = False
-            else:
-                return True
-
-        # show rendered text via urwid
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        return False
-
 # bad pattern: should be configurable
 def renderHeader(character):
     """
@@ -5251,7 +3082,7 @@ def renderHeader(character):
     return ""
 
     # render the sections to display
-    questSection = renderQuests(maxQuests=2)
+    questSection = QuestMenu.renderQuests(maxQuests=2)
     messagesSection = renderMessages(character)
 
     # calculate the size of the elements
@@ -5342,1238 +3173,6 @@ def renderMessages(character, maxMessages=5):
 
     return txt
 
-# bad code: the asList and questIndex parameters are out of place
-def renderQuests(maxQuests=0, char=None, asList=False, questCursor=None,sidebared=False):
-    """
-    render the quests into a string or list
-
-    Parameters:
-        maxQuests: the maximal number of quests rendered (0=no maximum)
-        char: the character the quests belong to
-        asList: flag to switch from rendering as string to rendering as list
-        questsIndex: index pointing to the active quest
-
-    Returns:
-        the rendered messages
-    """
-
-    # basic set up
-    if not char:
-        char = src.gamestate.gamestate.mainChar
-    if asList:
-        txt = []
-    else:
-        txt = ""
-
-    # render the quests
-    if len(char.quests):
-        if sidebared:
-            result = char.quests[0].getSolvingCommandString(char)
-            solvingCommangString = None
-            if result:
-                if isinstance(result,list):
-                    result = (result,"continue")
-                if isinstance(result,str):
-                    result = (result,"continue")
-                (solvingCommangString,reason) = result
-                if isinstance(solvingCommangString,list):
-                    solvingCommangString = "".join(solvingCommangString)
-                if solvingCommangString:
-                    solvingCommangString = solvingCommangString.replace("\n","\\n")
-
-            if solvingCommangString:
-                nextstep = f"suggested action: \npress {solvingCommangString} \nto {reason}\n\n"
-            else:
-                nextstep = "suggested action: \npress + \nto generate subquests\n\n"
-            txt.append(src.interaction.ActionMeta(payload="+",content=nextstep))
-
-        if not sidebared:
-            baseList = char.quests
-            for index in questCursor:
-                quest = baseList[index]
-                try:
-                    baseList = quest.subQuests
-                except:
-                    baseList = None
-            txt.append(quest.generateTextDescription())
-            txt.append("\n")
-            txt.append("\n")
-
-            solvingCommangString = char.getActiveQuest().getSolvingCommandString(char)
-
-        if not sidebared:
-            txt.append("select quest:\n\n")
-
-        counter = 0
-        for quest in char.quests:
-            if questCursor and counter == questCursor[0]:
-                newCursor = questCursor[1:]
-            else:
-                newCursor = None
-            txt.extend(
-                quest.render(cursor=newCursor,sidebared=sidebared)
-                    )
-            txt.extend("\n\n")
-            counter += 1
-
-
-        if sidebared:
-            txt.append("press q to see detailed descriptions\n\n")
-    else:
-        txt.append("No Quest\n\n")
-
-    return txt
-
-
-#bad code: global function
-#bad code: should be abstracted
-def renderInventory(character=None,sidebared=False,cursor=None):
-    """
-    render the inventory of the player into a string
-
-    Returns:
-        the rendered string
-    """
-
-    if character is None:
-        char = src.gamestate.gamestate.mainChar
-    else:
-        char = character
-
-    txt = []
-    if not sidebared:
-        txt.append("your inventory:\n\n")
-    if len(char.inventory):
-        counter = 0
-        for item in char.inventory:
-            counter += 1
-            if not sidebared and counter == cursor+1:
-                txt.extend([ "-> "])
-            if isinstance(item.render(), int):
-                txt.extend(
-                    [
-                        str(counter),
-                        " - ",
-                        src.canvas.displayChars.indexedMapping[item.render()],
-                        " - ",
-                        item.name,
-                        "\n"])
-                if not sidebared and counter == cursor+1:
-                    txt.extend([
-                            item.getDetailedInfo(),
-                            "\n\n",
-                        ]
-                    )
-            else:
-                txt.extend(
-                    [
-                        str(counter),
-                        " - ",
-                        item.render(),
-                        " - ",
-                        item.name,
-                        "\n"])
-                if not sidebared and counter == cursor+1:
-                    txt.extend([
-                        item.getDetailedInfo(),
-                        "\n\n",
-                        ]
-                    )
-        txt.append("\n")
-        if not sidebared:
-            txt.append("press ws to move cursor\npress L to drop item nearby\npress l to drop item\n")
-    else:
-        txt.append("empty Inventory\n\n")
-    return txt
-
-# bad code: uses global function to render
-class HelpMenu(SubMenu):
-    """
-    the help submenue
-    """
-
-    type = "HelpMenu"
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the help text and ignore keypresses
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key in ("esc"," "):
-            character.changed("closedHelp")
-            return True
-
-        # show info
-        header.set_text((urwid.AttrSpec("default", "default"), "\n\nhelp\n\n"))
-        self.persistentText = ""
-        self.persistentText += renderHelp()
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        return False
-
-class ViewNPCsMenu(SubMenu):
-    type = "ViewNPCsMenu"
-
-    def __init__(self,personnelArtwork):
-        super().__init__()
-        self.index = 0
-        self.personnelArtwork = personnelArtwork
-        self.lastSelectedCharacter = None
-
-    def handleKey(self, key, noRender=False, character = None):
-
-        # exit the submenu
-        if key in ("esc"," ",):
-            return True
-
-        self.persistentText = ["press w/a and s/d to scroll\n\n"]
-        self.persistentText = ["press . to pass a turn\n\n"]
-        self.persistentText = ["press t to take over clone\n\n"]
-        self.persistentText = ["press r to reset clone quests\n\n"]
-
-        characters = self.personnelArtwork.getPersonnelList()
-
-        if not characters:
-            main.set_text((urwid.AttrSpec("default", "default"), "no personnel found"))
-            return None
-
-        if key in (".",):
-            character.timeTaken += 1
-
-        if key in ("w","a",):
-            if self.index > 0:
-                self.index -= 1
-            else:
-                self.index = len(characters)-1
-            self.lastSelectedCharacter = None
-
-        if key in ("s","d"):
-            if self.index < len(characters)-1:
-                self.index += 1
-            else:
-                self.index = 0
-            self.lastSelectedCharacter = None
-
-        self.persistentText.append(f"{self.index+1} of {len(characters)}\n\n")
-
-        if self.lastSelectedCharacter:
-            counter = 0
-            for char in characters:
-                if char == self.lastSelectedCharacter:
-                    self.index = counter
-                    break
-                counter += 1
-
-        selectedCharacter = characters[self.index]
-        self.lastSelectedCharacter = selectedCharacter
-
-        if key in ("t",):
-            src.gamestate.gamestate.mainChar = selectedCharacter
-        if key in ("r",):
-            for quest in selectedCharacter.quests:
-                quest.fail()
-
-            containerQuest = src.quests.questMap["BeUsefull"]()
-            selectedCharacter.quests.append(containerQuest)
-            containerQuest.assignToCharacter(selectedCharacter)
-            containerQuest.activate()
-            containerQuest.autoSolve = True
-
-            selectedCharacter.timeTaken = 0
-
-        self.persistentText.append("\nname: {} (marked by {})".format(selectedCharacter.name,"XX"))
-        part1 = f"position: {selectedCharacter.getPosition()} "
-        part2 = f"big position: {selectedCharacter.getBigPosition()} "
-        self.persistentText.append(f"\n{part1}{part2}")
-
-        self.persistentText.append(" "*40+"\n")
-        self.persistentText.append("\n")
-        if selectedCharacter.container.isRoom:
-            pos = selectedCharacter.getBigPosition()
-            smallPos = selectedCharacter.getPosition()
-            rawRender = selectedCharacter.container.render()
-            terrain = selectedCharacter.getTerrain()
-            miniMapRender = terrain.renderTiles()
-
-            y = 0
-            self.persistentText.append("\n")
-            while y < 15:
-                if y == 0 or y == 14:
-                    self.persistentText.append("  "*15)
-                else:
-                    x = 0
-                    self.persistentText.append("  ")
-                    for entry in rawRender[y-1]:
-                        if (x,y-1,0) == smallPos:
-                            self.persistentText.append("XX")
-                        else:
-                            self.persistentText.append(entry)
-                        x += 1
-                    self.persistentText.append("  ")
-                self.persistentText.append("  |  ")
-                x = 0
-                for entry in miniMapRender[y]:
-                    if (x,y,0) == pos:
-                        self.persistentText.append("XX")
-                    else:
-                        self.persistentText.append(entry)
-
-                    x += 1
-                self.persistentText.append("\n")
-                y += 1
-            self.persistentText.append("\n")
-        else:
-            pos = selectedCharacter.getBigPosition()
-            fullPos = selectedCharacter.getPosition()
-            rawRender = selectedCharacter.container.render(coordinateOffset=(15*pos[1],15*pos[0]),size=(14,14))
-            terrain = selectedCharacter.getTerrain()
-            miniMapRender = terrain.renderTiles()
-
-            y = 0
-            for line in rawRender:
-                x = 0
-                for entry in line:
-                    if (x+pos[0]*15,y+pos[1]*15,0) == fullPos:
-                        self.persistentText.append("XX")
-                    else:
-                        self.persistentText.append(entry)
-                    x += 1
-                self.persistentText.append("  |  ")
-                x = 0
-                for entry in miniMapRender[y]:
-                    if (x,y,0) == pos:
-                        self.persistentText.append("XX")
-                    else:
-                        self.persistentText.append(entry)
-
-                    x += 1
-                self.persistentText.append("\n")
-                y += 1
-        self.persistentText.append("\nrank: %s"%(selectedCharacter.rank))
-        self.persistentText.append("\ninventory: ")
-        for item in selectedCharacter.inventory:
-            self.persistentText.append(item.render())
-            self.persistentText.append(" ")
-        self.persistentText.append(f"({len(selectedCharacter.inventory)})")
-
-        if selectedCharacter.weapon:
-            self.persistentText.append("\nweapon: %s"%(selectedCharacter.weapon.baseDamage))
-        else:
-            self.persistentText.append("\nweapon: None")
-        if selectedCharacter.armor:
-            self.persistentText.append("\narmor: %s"%(selectedCharacter.armor.armorValue))
-        else:
-            self.persistentText.append("\narmor: None")
-        self.persistentText.append("\nstaff: %s"%(selectedCharacter.isStaff))
-        self.persistentText.append("\nduties: {}".format(", ".join(selectedCharacter.duties)))
-        quest = selectedCharacter.getActiveQuest()
-        if quest:
-            self.persistentText.append(f"\nactive quest: {quest.description}")
-        else:
-            self.persistentText.append("\nactive quest: None")
-
-        main.set_text(self.persistentText)
-        return None
-
-class StaffAsMatrixMenu(SubMenu):
-    type = "StaffAsMatrixMenu"
-
-    def __init__(self,staffArtwork):
-        super().__init__()
-        self.index = [0,0]
-        self.staffArtwork = staffArtwork
-        self.roomTypes = ["TrapRoom","WorkshopRoom"]
-
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the help text and ignore keypresses
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key in ("esc"," ",):
-            return True
-
-        text = ["press wasd to move cursor\npress j to increase\npress k to decrease\n"]
-
-        if key in ("w",) and self.index[1] > 0:
-            self.index[1] -= 1
-        if key in ("s",) and self.index[1] < len(self.roomTypes)-1:
-            self.index[1] += 1
-        if key in ("a",) and self.index[0] > 0:
-            self.index[0] -= 1
-        if key in ("d",):
-            self.index[0] += 1
-
-        if key in ("j",):
-            roomCounter = 1
-            for room in self.staffArtwork.container.container.rooms:
-                if room.objType == self.roomTypes[self.index[1]]:
-                    if self.index[0] == 0 or self.index[0] == roomCounter:
-                        self.staffArtwork.autoFillStaffFromMap({"character":character,"coordinate":(room.xPosition,room.yPosition)},redirect=False)
-                    roomCounter += 1
-        if key in ("k",):
-            roomCounter = 1
-            for room in self.staffArtwork.container.container.rooms:
-                if room.objType == self.roomTypes[self.index[1]]:
-                    if self.index[0] == 0 or self.index[0] == roomCounter:
-                        self.staffArtwork.autoRemoveStaffFromMap({"character":character,"coordinate":(room.xPosition,room.yPosition)},redirect=False)
-                    roomCounter += 1
-
-        counter = 0
-        for roomType in self.roomTypes:
-            color = "#fff"
-            if counter == self.index[1] and self.index[0] == 0:
-                color = "#f00"
-            text.append((urwid.AttrSpec(color, "default"),f"{roomType}"))
-            roomCounter = 1
-            for room in self.staffArtwork.container.container.rooms:
-                if room.objType == roomType:
-                    color = "#fff"
-                    if counter == self.index[1] and roomCounter == self.index[0]:
-                        color = "#f00"
-                    text.append((urwid.AttrSpec(color, "default")," %s"%(len(room.staff))))
-                    roomCounter += 1
-            text.append("\n")
-            counter += 1
-
-        # show info
-        header.set_text((urwid.AttrSpec("default", "default"), "\n\nhelp\n\n"))
-        self.persistentText = text
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        return False
-
-    def getQuestMarkersTile(self, character):
-        out = []
-
-        roomType = self.roomTypes[self.index[1]]
-        roomCounter = 1
-        for room in self.staffArtwork.container.container.rooms:
-            if room.objType != roomType:
-                continue
-            if self.index[0] == 0 or self.index[0] == roomCounter:
-                out.append((room.getPosition(),"selected"))
-            roomCounter += 1
-        return out
-
-
-class JobAsMatrixMenu(SubMenu):
-    type = "JobAsMatrixMenu"
-
-    def __init__(self,dutyArtwork):
-        super().__init__()
-        self.dutyArtwork = dutyArtwork
-        self.index = [0,0]
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the help text and ignore keypresses
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key in ("esc"," ",):
-            return True
-
-        terrain = self.dutyArtwork.getTerrain()
-        npcs = []
-        for char in terrain.characters:
-            if char.burnedIn:
-                continue
-            if char not in npcs:
-                npcs.append(char)
-        for room in terrain.rooms:
-            for char in room.characters:
-                if char.burnedIn:
-                    continue
-                if char not in npcs:
-                    npcs.append(char)
-        if src.gamestate.gamestate.mainChar in npcs:
-            npcs.remove(src.gamestate.gamestate.mainChar)
-
-        duties = list(reversed(["manufacturing","epoch questing","scavenging","machine operation","clone spawning","city planning","cleaning","painting","maggot gathering","machine placing","room building","machining","metal working","hauling","resource fetching","scrap hammering","resource gathering","questing","flask filling","praying"]))
-
-
-        if key == "C":
-            for npc in npcs:
-                npc.duties = []
-        if key == "w" and not self.index[0] < 1:
-            self.index[0] -= 1
-        if key == "s":
-            self.index[0] += 1
-        if key == "a" and not self.index[1] < 1:
-            self.index[1] -= 1
-        if key == "d" and not self.index[1] > len(duties)-2:
-            self.index[1] += 1
-        if key in ("j","k","l","c"):
-            rowCounter = 0
-            for npc in npcs:
-                if npc.faction != character.faction:
-                    continue
-                if isinstance(npc,src.characters.Ghoul):
-                    continue
-
-                if rowCounter == self.index[0]:
-                    dutyname = duties[self.index[1]]
-
-                    if key == "l":
-                        pass
-                    elif key == "j":
-                        if not dutyname in npc.duties:
-                            npc.duties.append(dutyname)
-                            npc.dutyPriorities[dutyname] = 1
-                        else:
-                            if not dutyname in npc.dutyPriorities:
-                                npc.dutyPriorities[dutyname] = 1
-                            npc.dutyPriorities[dutyname] += 1
-                    elif key == "k":
-                        if dutyname in npc.duties:
-                            if not dutyname in npc.dutyPriorities:
-                                npc.dutyPriorities[dutyname] = 1
-                            npc.dutyPriorities[dutyname] -= 1
-                            if npc.dutyPriorities[dutyname] < 1:
-                                del npc.dutyPriorities[dutyname]
-                                npc.duties.remove(dutyname)
-                    elif key == "c":
-                        npc.duties = []
-                        npc.dutyPriorities = {}
-
-                rowCounter += 1
-
-        text = "press wasd to move cursor"
-        text += "press j to enable/disable"
-
-        text = [text]
-
-        text.append("\ncharacter                 ")
-        rowCounter = 0
-        dutyCounter = 0
-        for duty in duties:
-            color = "default"
-            if rowCounter == self.index[1]:
-                color = "#555"
-            text.append("|")
-            text.append((urwid.AttrSpec("default", color)," "+duty+" "))
-            rowCounter += 1
-            dutyCounter += 1
-            if dutyCounter == 6:
-                text.append("\n                          ")
-                dutyCounter = 0
-
-
-        def convertName(name):
-            return name.ljust(25," ")[0:24]
-
-
-        lineCounter = 0
-        color = "default"
-        rowCounter = 0
-        if lineCounter == self.index[0]:
-            color = "#333"
-
-        lineCounter = 0
-        for npc in npcs:
-            if npc.faction != character.faction:
-                continue
-            if isinstance(npc,src.characters.Ghoul):
-                continue
-            text.append("\n")
-            if lineCounter == self.index[0]:
-                color = "#333"
-            else:
-                color = "default"
-            text.append((urwid.AttrSpec("default", color),f"{convertName(npc.name)}: "))
-            rowCounter = 0
-            for duty in duties:
-                if lineCounter == self.index[0] and rowCounter == self.index[1]:
-                    text.append("=>")
-                else:
-                    color = "default"
-                    if rowCounter == self.index[1] or lineCounter == self.index[0]:
-                        color = "#333"
-                    text.append((urwid.AttrSpec("default", color),"  "))
-
-                if duty in npc.duties:
-                    text.append(str(npc.dutyPriorities.get(duty,1)))
-                else:
-                    color = "default"
-                    if rowCounter == self.index[1] or lineCounter == self.index[0]:
-                        color = "#333"
-                    text.append((urwid.AttrSpec("default", color)," "))
-                text.append("|")
-                rowCounter += 1
-            lineCounter += 1
-
-        # show info
-        header.set_text((urwid.AttrSpec("default", "default"), "\n\nhelp\n\n"))
-        self.persistentText = text
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        return False
-
-class MapMenu(SubMenu):
-    """
-    a menu for triggering actions from a map
-    """
-
-    type = "MapMenu"
-
-    def __init__(self, mapContent=None,functionMap=None, extraText = "", cursor = None):
-        """
-        initialise internal state
-
-        Parameters:
-            mapContent: the content to show
-        """
-
-        super().__init__()
-        self.mapContent = mapContent
-        self.functionMap = functionMap
-        self.extraText = extraText
-        if cursor:
-            self.cursor = (cursor[0],cursor[1],)
-        else:
-            self.cursor = (7,7)
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the map and trigger functions depending on key presses
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        closeMenu = False
-        mappedFunctions = self.functionMap.get(self.cursor, {})
-        if key in mappedFunctions:
-            closeMenu = True
-            self.callIndirect(mappedFunctions[key]["function"],{"coordinate":self.cursor})
-
-        # exit the submenu
-        if key in ("w",) and self.cursor[1] > 1:
-            self.cursor = (self.cursor[0],self.cursor[1]-1)
-        if key in ("s",) and self.cursor[1] < 13:
-            self.cursor = (self.cursor[0],self.cursor[1]+1)
-        if key in ("a",) and self.cursor[0] > 1:
-            self.cursor = (self.cursor[0]-1,self.cursor[1])
-        if key in ("d",) and self.cursor[0] < 13:
-            self.cursor = (self.cursor[0]+1,self.cursor[1])
-
-        if closeMenu or key in (
-            "esc",
-            "enter",
-            "space",
-            "j",
-        ):
-            if self.followUp:
-                self.followUp()
-            return True
-
-        quest = character.getActiveQuest()
-        if quest:
-            for marker in quest.getQuestMarkersTile(character):
-                pos = marker[0]
-                display = self.mapContent[pos[1]][pos[0]]
-
-                actionMeta = None
-                if isinstance(display,src.interaction.ActionMeta):
-                    actionMeta = display
-                    display = display.content
-
-                if isinstance(display,int):
-                    display = src.canvas.displayChars.indexedMapping[display]
-                if isinstance(display,str):
-                    display = (src.interaction.urwid.AttrSpec("#fff","black"),display)
-
-                if hasattr(display[0],"fg"):
-                    display = (src.interaction.urwid.AttrSpec(display[0].fg,"#555"),display[1])
-                else:
-                    if not isinstance(display[0],tuple):
-                        display = (src.interaction.urwid.AttrSpec(display[0].foreground,"#555"),display[1])
-
-                if actionMeta:
-                    actionMeta.content = display
-                    display = actionMeta
-
-                self.mapContent[pos[1]][pos[0]] = display
-
-        # show rendered map
-        mapText = []
-        for y in range(15):
-            mapText.append([])
-            for x in range(15):
-                if (x,y) == self.cursor:
-                    mapText[-1].append("██")
-                else:
-                    mapText[-1].append(self.mapContent[y][x])
-            mapText[-1].append("\n")
-
-        mapText.append(f"\n press wasd to move cursor {self.cursor}")
-
-        mappedFunctions = self.functionMap.get(self.cursor, {})
-        for (key,item) in mappedFunctions.items():
-            mapText.append("\n press {} to {}".format(key,item["description"],))
-
-        mapText.append(self.extraText)
-
-        if not noRender:
-            # show info
-            header.set_text((urwid.AttrSpec("default", "default"), ""))
-            self.persistentText = mapText
-            main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-
-        return False
-
-class TextMenu(SubMenu):
-    """
-    a menu showing a text
-    """
-
-    type = "TextMenu"
-
-    def __init__(self, text=""):
-        """
-        initialise internal state
-
-        Parameters:
-            text: the text to show
-        """
-
-        super().__init__()
-        self.text = text
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the text and ignore keypresses
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key in (
-            "esc",
-            "enter",
-            "space",
-            " ",
-        ):
-            if self.followUp:
-                self.callIndirect(self.followUp)
-            return True
-
-        if not noRender:
-            # show info
-            header.set_text((urwid.AttrSpec("default", "default"), ""))
-            self.persistentText = self.text
-            main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        return False
-
-class OneKeystrokeMenu(SubMenu):
-    """
-    a menu getting a single keystroke from the character
-    """
-
-    type = "OneKeystrokeMenu"
-
-    def __init__(self, text="",targetParamName="keyPressed"):
-        """
-        initialise inernal state
-
-        Parameters:
-            text: the text to show
-        """
-
-        super().__init__()
-        self.text = text
-        self.firstRun = True
-        self.keyPressed = ""
-        self.done = False
-        self.targetParamName = targetParamName
-        self.counter = 0
-        self.rerenderFunction = None
-
-    def rerender(self):
-        if self.rerenderFunction:
-            main.set_text((urwid.AttrSpec("default", "default"), self.rerenderFunction()))
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the text and quit on second keypress
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # show info
-        if not noRender:
-            header.set_text((urwid.AttrSpec("default", "default"), ""))
-            self.persistentText = self.text
-            main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        # exit the submenu
-        if key not in ("~",) and not self.firstRun:
-            self.keyPressed = key
-            if self.followUp:
-                self.callIndirect(self.followUp,{self.targetParamName:key})
-            self.done = True
-            return True
-
-        self.firstRun = False
-
-        return False
-
-class ImplantConnection(SubMenu):
-    type = "ImplantConnection"
-
-    def __init__(self, connectionTarget):
-        super().__init__()
-        self.connectionTarget = connectionTarget
-        self.submenu = None
-        self.sidebared = False
-
-    def handleKey(self, key, noRender=False, character = None):
-        if not noRender:
-            header.set_text((urwid.AttrSpec("default", "default"), ""))
-            self.persistentText = f"implant connection to {self.connectionTarget.type}"
-            self.persistentText += "\n\n"
-            self.persistentText += "press j to use connection\n"
-            self.persistentText += "press x to close connection\n"
-            main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        if self.submenu and self.submenu.done:
-            self.submenu = None
-            return False
-
-        if self.submenu:
-            self.submenu.handleKey(key, noRender, character)
-            return False
-
-        if key in ("x",):
-            self.done = True
-            return True
-        if character and key in ("ESC","lESC",):
-            character.rememberedMenu.append(self)
-            return True
-        if character and key in ("rESC",):
-            character.rememberedMenu2.append(self)
-            return True
-
-        if character and key in ("j",):
-            self.connectionTarget.apply(character)
-            character.rememberedMenu.append(self)
-
-        return False
-
-    def render(self, char):
-        return f"implant connection to {self.connectionTarget.type}"
-
-class ChangeViewsMenu(SubMenu):
-    type = "ChangeViewsMenu"
-
-    def __init__(self):
-        super().__init__()
-        self.firstKey = True
-
-    def handleKey(self, key, noRender=False, character = None):
-
-        if character and key in ("a",):
-            character.personality["viewChar"] = "activity"
-        if character and key in ("A",):
-            character.personality["viewColour"] = "activity"
-        if character and key in ("r",):
-            character.personality["viewChar"] = "rank"
-        if character and key in ("R",):
-            character.personality["viewColour"] = "rank"
-        if character and key in ("h",):
-            character.personality["viewChar"] = "health"
-        if character and key in ("H",):
-            character.personality["viewColour"] = "health"
-        if character and key in ("n",):
-            character.personality["viewChar"] = "name"
-        if character and key in ("N",):
-            character.personality["viewColour"] = "name"
-        if character and key in ("f",):
-            character.personality["viewChar"] = "faction"
-        if character and key in ("F",):
-            character.personality["viewColour"] = "faction"
-
-        viewChar = character.personality["viewChar"]
-        viewColour = character.personality["viewColour"]
-
-        self.persistentText = []
-        self.persistentText.append("change view menu\n\n")
-        color = "#fff"
-        if viewChar == "activity":
-            color = "#f00"
-        if viewColour == "activity":
-            color = "#0f0"
-        self.persistentText.append((urwid.AttrSpec(color, "default"),"press a/A to show NPC activity marking\n"))
-
-        color = "#fff"
-        if viewChar == "rank":
-            color = "#f00"
-        if viewColour == "rank":
-            color = "#0f0"
-        self.persistentText.append((urwid.AttrSpec(color, "default"),"press r/R to rank marking\n"))
-        color = "#fff"
-        if viewChar == "health":
-            color = "#f00"
-        if viewColour == "health":
-            color = "#0f0"
-        self.persistentText.append((urwid.AttrSpec(color, "default"),"press h/H to show health marking\n"))
-        color = "#fff"
-        if viewChar == "name":
-            color = "#f00"
-        if viewColour == "name":
-            color = "#0f0"
-        self.persistentText.append((urwid.AttrSpec(color, "default"),"press n/N to show name marking\n"))
-        color = "#fff"
-        if viewChar == "faction":
-            color = "#f00"
-        if viewColour == "faction":
-            color = "#0f0"
-        self.persistentText.append((urwid.AttrSpec(color, "default"),"press f/F to show faction indicator\n"))
-        self.persistentText.append("\n\nsmall letters for display color, big letters for display char")
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        # exit the submenu
-        if key in ("esc",):
-            self.done = True
-            return True
-        return None
-
-class RoomMenu(SubMenu):
-    type = "RoomMenu"
-
-    def __init__(self, room):
-        super().__init__()
-        self.room = room
-        self.submenu = None
-        self.firstKey = True
-
-    def handleKey(self, key, noRender=False, character = None):
-        self.persistentText = "room menu \n\n"
-
-        self.persistentText = [self.persistentText]
-        self.persistentText.append(f"{self.room.objType} - {self.room.tag}\n")
-        if hasattr(self.room,"chargeStrength"):
-            self.persistentText.append("chargeStrength: " + str(self.room.chargeStrength)+"\n")
-        if hasattr(self.room,"electricalCharges"):
-            self.persistentText.append("electricalCharges: " + str(self.room.electricalCharges)+"\n")
-        if hasattr(self.room,"maxElectricalCharges"):
-            self.persistentText.append("maxElectricalCharges: " + str(self.room.maxElectricalCharges)+"\n")
-        self.persistentText.append("\n\n")
-        if self.room.staff:
-            self.persistentText.append("staff:\n")
-            for staffNpc in self.room.staff:
-                deadText = ""
-                if staffNpc.dead:
-                    deadText = " (dead)"
-                questText = ""
-                if not staffNpc.dead and staffNpc.quests:
-                    questText = staffNpc.quests[0].description.split("\n")[0]
-                    try:
-                        questText += staffNpc.quests[0].description.split("\n")[1]
-                    except:
-                        pass
-                self.persistentText.append(f"{staffNpc.name}{deadText} - {questText}\n")
-        else:
-                self.persistentText.append("There is no staff assigned.\nassign staff by using the staff artwork (SA)")
-
-        self.persistentText.append(f"\n\nRoompriority: {self.room.priority}")
-
-        if self.room.floorPlan:
-            self.persistentText.append("\n\nThis room has a floor plan.")
-            if "walkingSpaces" in self.room.floorPlan:
-                logger.info("walkingSpaces")
-                logger.info(self.room.floorPlan["walkingSpaces"])
-            if "buildSites" in self.room.floorPlan:
-                logger.info("buildSites")
-                logger.info(self.room.floorPlan["buildSites"])
-            if "storageSlots" in self.room.floorPlan:
-                logger.info("storageSlots")
-                logger.info(self.room.floorPlan["storageSlots"])
-
-        try:
-            self.room.requiredDuties
-        except:
-            self.room.requiredDuties = []
-
-        if self.room.requiredDuties:
-            self.persistentText.append("\n\nThis room has required duties.\n%s"%self.room.requiredDuties)
-
-        self.persistentText.append("\n\n- q: open staff section\n- r: show resource sources\n- o: issue room orders")
-
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        if self.firstKey:
-            self.firstKey = False
-            return None
-
-        if character and key in ("q",):
-            character.macroState["submenue"] = RoomDutyMenu(self.room)
-
-        if character and key in ("r",):
-            character.macroState["submenue"] = RoomSourceMenu(self.room)
-
-        if character and key in ("o",):
-            homeRoom = character.getHomeRoom()
-            items = homeRoom.getItemsByType("OrderArtwork",needsBolted=True)
-            if not items:
-                character.addMessage("order artwork not found")
-                return True
-            item = items[0]
-            item.showMap(character)
-            self.done = True
-            return True
-
-        # exit the submenu
-        if key in ("esc",):
-            self.done = True
-            return True
-        return None
-
-class RoomSourceMenu(SubMenu):
-    """
-    """
-
-    type = "RoomSourceMenu"
-    def __init__(self, room):
-        """
-        initialise inernal state
-
-        Parameters:
-            text: the text to show
-        """
-
-        super().__init__()
-        self.room = room
-        self.submenu = None
-
-    def handleKey(self, key, noRender=False, character = None):
-        if self.submenu and self.submenu.done:
-            sourceType = self.submenu.text.split(":")[0].strip()
-            rawCoordinate = self.submenu.text.split(":")[1].strip(" ()")
-            coordinate = []
-            for num in rawCoordinate.split(","):
-                coordinate.append(int(num))
-            while len(coordinate) > 2:
-                coordinate.pop()
-            coordinate = tuple(coordinate)
-            source = (coordinate,sourceType)
-            character.container.sources.append(source)
-
-            self.submenu = None
-
-
-        if self.submenu:
-            self.submenu.handleKey(key, noRender, character)
-            if self.submenu.done:
-                self.handleKey("~",noRender,character)
-            return False
-
-        self.persistentText = "sources to fetch resources from:\n\n"
-        for source in self.room.sources:
-            self.persistentText += f"{source[1]}: {source[0]}\n"
-        self.persistentText += "\n\npresss c to add source"
-        main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        if key == "c":
-            self.submenu = InputMenu("input source.\nCurrent tile is {}.\nFormat to input source is \nresourceType: tilecoordinate".format(character.container.getTilePosition()))
-            self.submenu.handleKey("~", noRender, character)
-            return False
-
-        # exit the submenu
-        if key in ("esc",):
-            self.done = True
-            return True
-        return None
-
-class RoomDutyMenu(SubMenu):
-    """
-    """
-
-    type = "RoomDutyMenu"
-
-    def __init__(self, room):
-        """
-        initialise inernal state
-
-        Parameters:
-            text: the text to show
-        """
-
-        super().__init__()
-        self.room = room
-        self.keyPressed = ""
-        self.done = False
-        self.index = (0,0)
-
-    def handleKey(self, key, noRender=False, character = None):
-        """
-        show the text and quit on second keypress
-
-        Parameters:
-            key: the key pressed
-            noRender: flag to skip rendering
-        Returns:
-            returns True when done
-        """
-
-        # exit the submenu
-        if key in ("w",) and self.index[1] > 0:
-            self.index = (self.index[0],self.index[1]-1)
-        if key in ("s",) and self.index[1] < len(self.room.duties)-1:
-            self.index = (self.index[0],self.index[1]+1)
-        if key in ("a",) and self.index[0] > 0:
-            self.index = (self.index[0]-1,self.index[1])
-        if key in ("d",) and self.index[0] < len(self.room.staff)-1:
-            self.index = (self.index[0]+1,self.index[1])
-
-        if key in ("j","enter"):
-            duty = self.room.duties[self.index[1]]
-            staffCharacter = self.room.staff[self.index[0]]
-            if duty in staffCharacter.duties:
-                staffCharacter.duties.remove(duty)
-            else:
-                staffCharacter.duties.append(duty)
-
-        # show info
-        if not noRender:
-            header.set_text((urwid.AttrSpec("default", "default"), ""))
-            self.persistentText = """
-move the cursor using w/a/s/d.
-press j or enter to select
-
-"""
-            self.persistentText = [self.persistentText]
-            self.persistentText.append(f"{self.room.objType} - {self.room.name}\n")
-            try:
-                self.persistentText.append( "electricalCharges: " + str(self.room.electricalCharges)+"\n")
-            except:
-                pass
-            try:
-                self.persistentText.append("maxElectricalCharges: " + str(self.room.maxElectricalCharges)+"\n")
-            except:
-                pass
-            self.persistentText.append("\n\n")
-            if self.room.staff:
-                self.persistentText.append("staff:\n")
-                for staffNpc in self.room.staff:
-                    deadText = ""
-                    if staffNpc.dead:
-                        deadText = " (dead)"
-                    questText = ""
-                    if not staffNpc.dead and staffNpc.quests:
-                        questText = staffNpc.quests[0].description.split("\n")[0]
-                        try:
-                            questText += staffNpc.quests[0].description.split("\n")[1]
-                        except:
-                            pass
-                    self.persistentText.append(f"{staffNpc.name}{deadText} - {questText}\n")
-            else:
-                    self.persistentText.append("There is no staff assigned assign staff by using the staff artwork (SA)")
-
-            self.persistentText += "\n"
-
-            if self.room.staff:
-                rowCounter = 0
-                for duty in self.room.duties:
-                    self.persistentText.append( duty + " |")
-                    colCounter = 0
-                    for staffCharacter in self.room.staff:
-                        frontColor = "#fff"
-                        if duty in staffCharacter.duties:
-                            frontColor = "#0e0"
-                        backColor = "#000"
-                        if (colCounter,rowCounter) == self.index:
-                            backColor = "#444"
-                        self.persistentText.append((urwid.AttrSpec(frontColor, backColor),staffCharacter.name))
-                        self.persistentText.append(" |")
-                        colCounter += 1
-                    self.persistentText.append("\n")
-                    rowCounter += 1
-            main.set_text((urwid.AttrSpec("default", "default"), self.persistentText))
-
-        # exit the submenu
-        if key in ("esc",):
-            self.done = True
-            return True
-
-        self.firstRun = False
-
-        return False
-
-
-# bad code: should not be a global function
-def renderHelp():
-    """
-    renders the help text
-
-    Returns:
-        the rendered help text
-    """
-
-    char = src.gamestate.gamestate.mainChar
-    txt = ""
-    txt += "\n= movement =\n"
-    txt += " w/a/s/d: move north/east/south/west (up/left/down/right)\n"
-    txt += " W/A/S/D: dash north/east/south/west (up/left/down/right)\n"
-    txt += " .: wait\n"
-    txt += "\n= item interaction =\n"
-    txt += " j/J: activate items\n"
-    txt += " c/C: complex activate items\n"
-    txt += " k/K: pick up item\n"
-    txt += " l/L: drop item\n"
-    txt += " e/E: examine item\n"
-    txt += "\n"
-    txt += "lowercase keys work on the square you stand on or the last item you bumped into\n"
-    txt += "uppercase keys open a secondary menu\n"
-    txt += "\n= fighting =\n"
-    txt += " w/a/s/d: attack north/east/south/west (up/left/down/right)\n"
-    txt += " W/A/S/D: advanced attack north/east/south/west (up/left/down/right)\n"
-    txt += "\n= user interface =\n"
-    txt += " q: open quests\n"
-    txt += " Q: open advanced quest menu\n"
-    txt += " i: show inventory\n"
-    txt += " t: set information to render\n"
-    txt += " r: show room menu\n"
-    txt += "\n"
-    txt += "sadly the controls cannot be changed at the moment\n"
-    txt += "if you have issues with the character running into wall, tap the keys instead of holding them\n"
-    txt += "\n"
-    txt += ""
-    return txt
 
 
 lastTerrain = None
