+++ src/menuFolder/__init__.py
@@ -1,1 +1,30 @@
+import src.menuFolder.SubMenu
+import src.menuFolder.AdvancedQuestMenu
+import src.menuFolder.ChangeViewsMenu
+import src.menuFolder.CharacterInfoMenu
+import src.menuFolder.ChatPartnerselection
+import src.menuFolder.CombatInfoMenu
+import src.menuFolder.CreateQuestMenu
+import src.menuFolder.DebugMenu
+import src.menuFolder.HelpMenu
+import src.menuFolder.IdleChatNPCMenu
+import src.menuFolder.ImplantConnection
+import src.menuFolder.InputMenu
+import src.menuFolder.InstructNPCMenu
+import src.menuFolder.InstructSubordinatesMenu
+import src.menuFolder.InventoryMenu
+import src.menuFolder.JobAsMatrixMenu
+import src.menuFolder.ListActionMenu
+import src.menuFolder.MapMenu
+import src.menuFolder.NameGhoulMenu
+import src.menuFolder.OneKeystrokeMenu
+import src.menuFolder.QuestMenu
+import src.menuFolder.RoomDutyMenu
+import src.menuFolder.RoomMenu
+import src.menuFolder.RoomSourceMenu
+import src.menuFolder.SelectionMenu
+import src.menuFolder.setNPCDutiesMenu
 import src.menuFolder.SettingMenu
+import src.menuFolder.StaffAsMatrixMenu
+import src.menuFolder.TextMenu
+import src.menuFolder.ViewNPCsMenu
